Remove legacy gateway HTTP and WebSocket clients
