--- src/app-core/gateway/gateway.test.ts
@@ -1,1873 +1,0 @@
-import { describe, expect, test, vi } from "vitest";
-import { DEFAULT_GATEWAY_CONFIG, resolveGatewayConfig } from "./gatewayConfig";
-import {
-  checkGatewayHealth,
-  createGatewayApiClient,
-  resolveTsCoworkRuntimeRollout,
-} from "./gatewayHttpClient";
-import {
-  createGatewaySocketMessage,
-  flushGatewaySocketQueue,
-  normalizeGatewayFrame,
-  sendGatewaySocketJson,
-} from "./gatewayWebSocketClient";
-
-describe("gateway config", () => {
-  test("builds default local HTTP and WebSocket endpoints", () => {
-    expect(DEFAULT_GATEWAY_CONFIG.httpBaseUrl).toBe("http://127.0.0.1:18790");
-    expect(DEFAULT_GATEWAY_CONFIG.wsUrl).toBe("ws://127.0.0.1:18790/ws");
-  });
-
-  test("normalizes provided URLs without duplicate slashes", () => {
-    const config = resolveGatewayConfig({
-      httpBaseUrl: "http://localhost:18790/",
-      wsUrl: "ws://localhost:18790/ws",
-      requestTimeoutMs: 250,
-    });
-
-    expect(config.httpBaseUrl).toBe("http://localhost:18790");
-    expect(config.wsUrl).toBe("ws://localhost:18790/ws");
-    expect(config.requestTimeoutMs).toBe(250);
-  });
-
-  test("resolves TS Cowork runtime rollout from desktop config with TS-first defaults", () => {
-    expect(resolveTsCoworkRuntimeRollout({})).toEqual({
-      enabled: true,
-      readOnlySnapshot: true,
-      mutations: true,
-      scheduler: true,
-      swarm: true,
-    });
-
-    expect(resolveTsCoworkRuntimeRollout({
-      desktop: {
-        ts_cowork_runtime: {
-          enabled: true,
-          read_only_snapshot: true,
-          mutations: true,
-          scheduler: true,
-        },
-      },
-    })).toEqual({
-      enabled: true,
-      readOnlySnapshot: true,
-      mutations: true,
-      scheduler: true,
-      swarm: true,
-    });
-
-    expect(resolveTsCoworkRuntimeRollout({
-      desktop: {
-        tsCoworkRuntime: {
-          enabled: false,
-          scheduler: false,
-        },
-      },
-    })).toMatchObject({
-      enabled: false,
-      scheduler: false,
-    });
-  });
-});
-
-describe("gateway HTTP client", () => {
-  test("reports a reachable gateway when HTTP and WebSocket checks pass", async () => {
-    const fetchFn = vi.fn(async (url: RequestInfo | URL, _init?: RequestInit) => {
-      if (String(url).endsWith("/webui/bootstrap")) {
-        return new Response(JSON.stringify({ token: "token-1", ws_path: "/ws" }), { status: 200 });
-      }
-      return new Response(JSON.stringify({ ok: true }), { status: 200 });
-    });
-    const webSocketProbe = vi.fn(async () => ({ ok: true as const }));
-
-    const result = await checkGatewayHealth({
-      config: DEFAULT_GATEWAY_CONFIG,
-      fetchFn,
-      webSocketProbe,
-    });
-
-    expect(result.state).toBe("running");
-    expect(result.http.ok).toBe(true);
-    expect(result.webSocket.ok).toBe(true);
-    expect(fetchFn).toHaveBeenCalledWith(
-      "http://127.0.0.1:18790/webui/bootstrap",
-      expect.objectContaining({ signal: expect.any(AbortSignal) }),
-    );
-    expect(fetchFn).toHaveBeenCalledWith(
-      "http://127.0.0.1:18790/api/status",
-      expect.objectContaining({
-        headers: expect.objectContaining({ Authorization: "Bearer token-1" }),
-        signal: expect.any(AbortSignal),
-      }),
-    );
-    expect(webSocketProbe).toHaveBeenCalledWith(
-      "ws://127.0.0.1:18790/ws?token=token-1",
-      DEFAULT_GATEWAY_CONFIG.requestTimeoutMs,
-    );
-  });
-
-  test("keeps endpoint details when the gateway is offline", async () => {
-    const fetchFn = vi.fn(async () => {
-      throw new Error("connect ECONNREFUSED");
-    });
-
-    const result = await checkGatewayHealth({
-      config: DEFAULT_GATEWAY_CONFIG,
-      fetchFn,
-      webSocketProbe: async () => ({ ok: false as const, error: "not checked" }),
-    });
-
-    expect(result.state).toBe("offline");
-    expect(result.http.ok).toBe(false);
-    if (!result.http.ok) {
-      expect(result.http.error).toContain("ECONNREFUSED");
-    }
-    expect(result.httpBaseUrl).toBe(DEFAULT_GATEWAY_CONFIG.httpBaseUrl);
-    expect(result.wsUrl).toBe(DEFAULT_GATEWAY_CONFIG.wsUrl);
-  });
-
-  test("constructs shared route group requests", async () => {
-    const fetchFn = vi.fn(async (url: RequestInfo | URL, _init?: RequestInit) => {
-      if (String(url).endsWith("/webui/bootstrap")) {
-        return new Response(JSON.stringify({ token: "token-1" }), { status: 200 });
-      }
-      return new Response(JSON.stringify({ items: [] }), { status: 200 });
-    });
-    const client = createGatewayApiClient({
-      config: DEFAULT_GATEWAY_CONFIG,
-      fetchFn,
-    });
-
-    await client.sessions.list();
-    await client.sessions.messages("WebSocket:chat-1");
-    await client.workspace.file("docs/readme.md");
-    await client.workspace.putFile("docs/readme.md", {
-      content: "# Readme\n",
-      expected_updated_at: "2026-05-31T10:00:00+00:00",
-    });
-    await client.openAi.health();
-    await client.openAi.models();
-    await client.cowork.summary("cowork-1");
-    await client.skills.create({ name: "planner" });
-    await client.skills.update("planner/phase", { content: "# Updated" });
-    await client.skills.validate("planner/phase");
-    await client.skills.delete("planner/phase");
-    await client.cowork.sessions({ includeCompleted: true, originChatId: "chat/1" });
-    await client.cowork.create({ goal: "Ship desktop" });
-    await client.cowork.run("cowork/1", { run_until_idle: true });
-    await client.cowork.action("cowork/1", "pause");
-    await client.cowork.action("cowork/1", "resume");
-    await client.cowork.action("cowork/1", "emergency-stop");
-    await client.cowork.delete("cowork/1");
-    await client.cowork.message("cowork/1", { content: "Continue", recipient_ids: [] });
-    await client.cowork.taskAction("cowork/1", "task/1", "assign", { assigned_agent_id: "agent-1" });
-    await client.cowork.workUnitAction("cowork/1", "wu/1", "retry", { reason: "retry" });
-    await client.cowork.selectBranch("cowork/1", "branch/1");
-    await client.cowork.selectBranchResult("cowork/1", "branch/1", { result_id: "result-1" });
-    await client.cowork.mergeBranchResults("cowork/1", { branch_ids: ["a", "b"] });
-    await client.cowork.validateBlueprint({ blueprint: {} }, { preview: true });
-
-    expect(fetchFn.mock.calls.map((call) => String((call as unknown[])[0]))).toEqual([
-      "http://127.0.0.1:18790/webui/bootstrap",
-      "http://127.0.0.1:18790/api/sessions",
-      "http://127.0.0.1:18790/api/sessions/WebSocket%3Achat-1/messages",
-      "http://127.0.0.1:18790/api/workspace/files/docs%2Freadme.md",
-      "http://127.0.0.1:18790/api/workspace/files/docs%2Freadme.md",
-      "http://127.0.0.1:18790/health",
-      "http://127.0.0.1:18790/v1/models",
-      "http://127.0.0.1:18790/api/cowork/sessions/cowork-1/summary",
-      "http://127.0.0.1:18790/api/skills",
-      "http://127.0.0.1:18790/api/skills/planner%2Fphase",
-      "http://127.0.0.1:18790/api/skills/planner%2Fphase/validate",
-      "http://127.0.0.1:18790/api/skills/planner%2Fphase",
-      "http://127.0.0.1:18790/api/cowork/sessions?include_completed=true&origin_chat_id=chat%2F1",
-      "http://127.0.0.1:18790/api/cowork/sessions",
-      "http://127.0.0.1:18790/api/cowork/sessions/cowork%2F1/run",
-      "http://127.0.0.1:18790/api/cowork/sessions/cowork%2F1/pause",
-      "http://127.0.0.1:18790/api/cowork/sessions/cowork%2F1/resume",
-      "http://127.0.0.1:18790/api/cowork/sessions/cowork%2F1/emergency-stop",
-      "http://127.0.0.1:18790/api/cowork/sessions/cowork%2F1",
-      "http://127.0.0.1:18790/api/cowork/sessions/cowork%2F1/messages",
-      "http://127.0.0.1:18790/api/cowork/sessions/cowork%2F1/tasks/task%2F1/assign",
-      "http://127.0.0.1:18790/api/cowork/sessions/cowork%2F1/work-units/wu%2F1/retry",
-      "http://127.0.0.1:18790/api/cowork/sessions/cowork%2F1/branches/branch%2F1/select",
-      "http://127.0.0.1:18790/api/cowork/sessions/cowork%2F1/branches/branch%2F1/result/select-final",
-      "http://127.0.0.1:18790/api/cowork/sessions/cowork%2F1/branch-results/merge",
-      "http://127.0.0.1:18790/api/cowork/blueprints/preview",
-    ]);
-    expect(fetchFn.mock.calls[4][1]).toMatchObject({
-      method: "PUT",
-      headers: expect.objectContaining({
-        Authorization: "Bearer token-1",
-        "Content-Type": "application/json",
-      }),
-      body: JSON.stringify({
-        content: "# Readme\n",
-        expected_updated_at: "2026-05-31T10:00:00+00:00",
-      }),
-    });
-    expect(fetchFn.mock.calls[9][1]).toMatchObject({
-      method: "PATCH",
-      headers: expect.objectContaining({
-        Authorization: "Bearer token-1",
-        "Content-Type": "application/json",
-      }),
-      body: JSON.stringify({ content: "# Updated" }),
-    });
-    expect(fetchFn.mock.calls[10][1]).toMatchObject({ method: "POST" });
-    expect(fetchFn.mock.calls[11][1]).toMatchObject({ method: "DELETE" });
-    expect(fetchFn.mock.calls[13][1]).toMatchObject({
-      method: "POST",
-      body: JSON.stringify({ goal: "Ship desktop" }),
-    });
-    expect(fetchFn.mock.calls[25][1]).toMatchObject({
-      method: "POST",
-      body: JSON.stringify({ blueprint: {} }),
-    });
-    for (const call of fetchFn.mock.calls.slice(1)) {
-      expect((call[1] as RequestInit).headers).toMatchObject({
-        Authorization: "Bearer token-1",
-      });
-    }
-  });
-
-  test("constructs approval resolution requests", async () => {
-    const fetchFn = vi.fn(async (url: RequestInfo | URL, _init?: RequestInit) => {
-      if (String(url).endsWith("/webui/bootstrap")) {
-        return new Response(JSON.stringify({ token: "token-1" }), { status: 200 });
-      }
-      return new Response(JSON.stringify({ ok: true }), { status: 200 });
-    });
-    const client = createGatewayApiClient({
-      config: DEFAULT_GATEWAY_CONFIG,
-      fetchFn,
-    });
-
-    await client.tools.approveApproval("approval/1", {
-      session_key: "WebSocket:chat-1",
-      scope: "session",
-      auto_retry: true,
-    });
-    await client.tools.denyApproval("approval/1", {
-      session_key: "WebSocket:chat-1",
-      auto_retry: true,
-    });
-
-    expect(fetchFn.mock.calls.map((call) => String((call as unknown[])[0]))).toEqual([
-      "http://127.0.0.1:18790/webui/bootstrap",
-      "http://127.0.0.1:18790/api/approvals/approval%2F1/approve",
-      "http://127.0.0.1:18790/api/approvals/approval%2F1/deny",
-    ]);
-    expect(fetchFn.mock.calls[1][1]).toMatchObject({
-      method: "POST",
-      headers: expect.objectContaining({
-        Authorization: "Bearer token-1",
-        "Content-Type": "application/json",
-      }),
-      body: JSON.stringify({
-        session_key: "WebSocket:chat-1",
-        scope: "session",
-        auto_retry: true,
-      }),
-    });
-    expect(fetchFn.mock.calls[2][1]).toMatchObject({
-      method: "POST",
-      headers: expect.objectContaining({
-        Authorization: "Bearer token-1",
-        "Content-Type": "application/json",
-      }),
-      body: JSON.stringify({
-        session_key: "WebSocket:chat-1",
-        auto_retry: true,
-      }),
-    });
-  });
-
-  test("prefers native skills operations when available", async () => {
-    const fetchFn = vi.fn(async (url: RequestInfo | URL, _init?: RequestInit) => {
-      if (String(url).endsWith("/webui/bootstrap")) {
-        return new Response(JSON.stringify({ token: "token-1" }), { status: 200 });
-      }
-      return new Response(JSON.stringify({ ok: true }), { status: 200 });
-    });
-    const nativeSkills = {
-      list: vi.fn(async () => ({ skills: [{ name: "planner" }] })),
-      detail: vi.fn(async (name: string) => ({ name, content: "Plan." })),
-      create: vi.fn(async (body: unknown) => ({ created: true, body })),
-      update: vi.fn(async (name: string, body: unknown) => ({ updated: true, name, body })),
-      delete: vi.fn(async (name: string) => ({ deleted: true, name })),
-      validate: vi.fn(async (name: string) => ({ name, valid: true })),
-    };
-    const client = createGatewayApiClient({
-      config: DEFAULT_GATEWAY_CONFIG,
-      fetchFn,
-      nativeSkills,
-    });
-
-    await expect(client.skills.list()).resolves.toEqual({ skills: [{ name: "planner" }] });
-    await expect(client.skills.detail("planner/phase")).resolves.toEqual({ name: "planner/phase", content: "Plan." });
-    await expect(client.skills.create({ name: "planner" })).resolves.toEqual({ created: true, body: { name: "planner" } });
-    await expect(client.skills.update("planner/phase", { content: "# Updated" })).resolves.toEqual({
-      updated: true,
-      name: "planner/phase",
-      body: { content: "# Updated" },
-    });
-    await expect(client.skills.delete("planner/phase")).resolves.toEqual({ deleted: true, name: "planner/phase" });
-    await expect(client.skills.validate("planner/phase")).resolves.toEqual({ name: "planner/phase", valid: true });
-
-    expect(nativeSkills.list).toHaveBeenCalledTimes(1);
-    expect(nativeSkills.detail).toHaveBeenCalledWith("planner/phase");
-    expect(nativeSkills.create).toHaveBeenCalledWith({ name: "planner" });
-    expect(nativeSkills.update).toHaveBeenCalledWith("planner/phase", { content: "# Updated" });
-    expect(nativeSkills.delete).toHaveBeenCalledWith("planner/phase");
-    expect(nativeSkills.validate).toHaveBeenCalledWith("planner/phase");
-    expect(fetchFn).not.toHaveBeenCalled();
-  });
-
-  test("prefers native WebUI status route when available", async () => {
-    const fetchFn = vi.fn(async () => new Response(JSON.stringify({ gateway: true }), { status: 200 }));
-    const nativeWebui = {
-      route: vi.fn(async (request: { method: string; path: string; body?: unknown }) => ({
-        channels: { websocket: { enabled: true, running: true } },
-        provider: null,
-        model: null,
-        request,
-      })),
-    };
-    const client = createGatewayApiClient({
-      config: DEFAULT_GATEWAY_CONFIG,
-      fetchFn,
-      nativeWebui,
-    });
-
-    await expect(client.runtime.status()).resolves.toMatchObject({
-      channels: { websocket: { enabled: true, running: true } },
-    });
-    expect(nativeWebui.route).toHaveBeenCalledWith({ method: "GET", path: "/api/status" });
-    expect(fetchFn).not.toHaveBeenCalled();
-  });
-
-  test("prefers native WebUI tools route when available", async () => {
-    const fetchFn = vi.fn(async () => new Response(JSON.stringify({ gateway: true }), { status: 200 }));
-    const nativeWebui = {
-      route: vi.fn(async (request: { method: string; path: string; body?: unknown }) => ({
-        tools: [{ name: "shell", description: "Run commands" }],
-        request,
-      })),
-    };
-    const client = createGatewayApiClient({
-      config: DEFAULT_GATEWAY_CONFIG,
-      fetchFn,
-      nativeWebui,
-    });
-
-    await expect(client.tools.list()).resolves.toEqual({
-      tools: [{ name: "shell", description: "Run commands" }],
-      request: { method: "GET", path: "/api/tools" },
-    });
-    expect(nativeWebui.route).toHaveBeenCalledWith({ method: "GET", path: "/api/tools" });
-    expect(fetchFn).not.toHaveBeenCalled();
-  });
-
-  test("prefers native Rust config read and native WebUI config patch when available", async () => {
-    const fetchFn = vi.fn(async () => new Response(JSON.stringify({ gateway: true }), { status: 200 }));
-    const nativeConfig = {
-      get: vi.fn(async () => ({
-        agents: { defaults: { provider: "dashscope", model: "qwen-max" } },
-      })),
-    };
-    const nativeWebui = {
-      route: vi.fn(async (request: { method: string; path: string; body?: unknown }) => ({
-        config: { agents: { defaults: { provider: "openrouter", model: "openai/gpt-4o-mini" } } },
-        request,
-      })),
-    };
-    const client = createGatewayApiClient({
-      config: DEFAULT_GATEWAY_CONFIG,
-      fetchFn,
-      nativeConfig,
-      nativeWebui,
-    });
-
-    await expect(client.config.get()).resolves.toEqual({
-      agents: { defaults: { provider: "dashscope", model: "qwen-max" } },
-    });
-    await expect(client.config.patch({
-      agents: { defaults: { provider: "openrouter", model: "openai/gpt-4o-mini" } },
-    })).resolves.toEqual({
-      config: { agents: { defaults: { provider: "openrouter", model: "openai/gpt-4o-mini" } } },
-      request: {
-        method: "PATCH",
-        path: "/api/config",
-        body: { agents: { defaults: { provider: "openrouter", model: "openai/gpt-4o-mini" } } },
-      },
-    });
-    expect(nativeConfig.get).toHaveBeenCalledTimes(1);
-    expect(nativeWebui.route).toHaveBeenCalledWith({
-      method: "PATCH",
-      path: "/api/config",
-      body: { agents: { defaults: { provider: "openrouter", model: "openai/gpt-4o-mini" } } },
-    });
-    expect(fetchFn).not.toHaveBeenCalled();
-  });
-
-  test("prefers native WebUI providers route when available", async () => {
-    const fetchFn = vi.fn(async () => new Response(JSON.stringify({ gateway: true }), { status: 200 }));
-    const nativeWebui = {
-      route: vi.fn(async (request: { method: string; path: string; body?: unknown }) => ({
-        providers: [{ id: "dashscope", displayName: "DashScope", status: "ready" }],
-        request,
-      })),
-    };
-    const client = createGatewayApiClient({
-      config: DEFAULT_GATEWAY_CONFIG,
-      fetchFn,
-      nativeWebui,
-    });
-
-    await expect(client.config.providers()).resolves.toEqual({
-      providers: [{ id: "dashscope", displayName: "DashScope", status: "ready" }],
-      request: { method: "GET", path: "/api/providers" },
-    });
-    expect(nativeWebui.route).toHaveBeenCalledWith({ method: "GET", path: "/api/providers" });
-    expect(fetchFn).not.toHaveBeenCalled();
-  });
-
-  test("prefers native WebUI provider models route when available", async () => {
-    const fetchFn = vi.fn(async () => new Response(JSON.stringify({ gateway: true }), { status: 200 }));
-    const nativeWebui = {
-      route: vi.fn(async (request: { method: string; path: string; body?: unknown }) => ({
-        ok: true,
-        models: ["qwen-max"],
-        request,
-      })),
-    };
-    const client = createGatewayApiClient({
-      config: DEFAULT_GATEWAY_CONFIG,
-      fetchFn,
-      nativeWebui,
-    });
-
-    await expect(client.config.providerModels({
-      provider: "dashscope",
-      refresh_live: true,
-    })).resolves.toEqual({
-      ok: true,
-      models: ["qwen-max"],
-      request: {
-        method: "POST",
-        path: "/api/provider-models",
-        body: { provider: "dashscope", refresh_live: true },
-      },
-    });
-    expect(fetchFn).not.toHaveBeenCalled();
-  });
-
-  test("prefers native Rust session list when both native paths are available", async () => {
-    const fetchFn = vi.fn(async () => new Response(JSON.stringify({ gateway: true }), { status: 200 }));
-    const nativeWebui = {
-      route: vi.fn(async () => {
-        throw new Error("native WebUI session list route should not be used");
-      }),
-    };
-    const nativeSessions = {
-      list: vi.fn(async () => ({
-        items: [{ key: "websocket:chat-1", chat_id: "chat-1", title: "Native session" }],
-      })),
-      messages: vi.fn(),
-    };
-    const client = createGatewayApiClient({
-      config: DEFAULT_GATEWAY_CONFIG,
-      fetchFn,
-      nativeSessions,
-      nativeWebui,
-    });
-
-    await expect(client.sessions.list()).resolves.toMatchObject({
-      items: [{ key: "websocket:chat-1", chat_id: "chat-1", title: "Native session" }],
-    });
-    expect(nativeSessions.list).toHaveBeenCalledTimes(1);
-    expect(nativeWebui.route).not.toHaveBeenCalled();
-    expect(fetchFn).not.toHaveBeenCalled();
-  });
-
-  test("does not hide native Rust session list failures behind gateway bootstrap fallback", async () => {
-    const fetchFn = vi.fn(async () => {
-      throw new Error("gateway fallback should not be used");
-    });
-    const nativeSessions = {
-      list: vi.fn(async () => {
-        throw new Error("native session list failed");
-      }),
-      messages: vi.fn(),
-    };
-    const client = createGatewayApiClient({
-      config: DEFAULT_GATEWAY_CONFIG,
-      fetchFn,
-      nativeSessions,
-    });
-
-    await expect(client.sessions.list()).rejects.toThrow("native session list failed");
-    expect(fetchFn).not.toHaveBeenCalled();
-  });
-
-  test("prefers native Rust session messages when both native paths are available", async () => {
-    const fetchFn = vi.fn(async () => new Response(JSON.stringify({ gateway: true }), { status: 200 }));
-    const nativeWebui = {
-      route: vi.fn(async () => {
-        throw new Error("native WebUI session messages route should not be used");
-      }),
-    };
-    const nativeSessions = {
-      list: vi.fn(),
-      messages: vi.fn(async (key: string) => ({
-        key,
-        messages: [{ role: "user", content: "Native history" }],
-      })),
-    };
-    const client = createGatewayApiClient({
-      config: DEFAULT_GATEWAY_CONFIG,
-      fetchFn,
-      nativeSessions,
-      nativeWebui,
-    });
-
-    await expect(client.sessions.messages("websocket:chat-1")).resolves.toMatchObject({
-      key: "websocket:chat-1",
-      messages: [{ role: "user", content: "Native history" }],
-    });
-    expect(nativeSessions.messages).toHaveBeenCalledWith("websocket:chat-1");
-    expect(nativeWebui.route).not.toHaveBeenCalled();
-    expect(fetchFn).not.toHaveBeenCalled();
-  });
-
-  test("prefers backend-authored native session effective capabilities", async () => {
-    const fetchFn = vi.fn(async () => new Response(JSON.stringify({ gateway: true }), { status: 200 }));
-    const nativeWebui = {
-      route: vi.fn(async () => {
-        throw new Error("native WebUI effective capabilities route should not be used");
-      }),
-    };
-    const nativeSessions = {
-      list: vi.fn(),
-      messages: vi.fn(),
-      effectiveCapabilities: vi.fn(async (key: string) => ({
-        schemaVersion: "tinybot.effective_capabilities.v1",
-        sessionId: key,
-      })),
-    };
-    const client = createGatewayApiClient({
-      config: DEFAULT_GATEWAY_CONFIG,
-      fetchFn,
-      nativeSessions,
-      nativeWebui,
-    });
-
-    await expect(client.sessions.effectiveCapabilities("websocket:chat-1")).resolves.toMatchObject({
-      sessionId: "websocket:chat-1",
-    });
-    expect(nativeSessions.effectiveCapabilities).toHaveBeenCalledWith("websocket:chat-1");
-    expect(nativeWebui.route).not.toHaveBeenCalled();
-    expect(fetchFn).not.toHaveBeenCalled();
-  });
-
-  test("prefers native Rust session state mutations when both native paths are available", async () => {
-    const fetchFn = vi.fn(async () => new Response(JSON.stringify({ gateway: true }), { status: 200 }));
-    const nativeWebui = {
-      route: vi.fn(async () => {
-        throw new Error("native WebUI session state route should not be used");
-      }),
-    };
-    const nativeSessions = {
-      list: vi.fn(),
-      messages: vi.fn(),
-      delete: vi.fn(async (key: string) => ({ key, deleted: true })),
-      patch: vi.fn(async (key: string, body: unknown) => ({ key, metadata: { pinned: true }, body })),
-      clear: vi.fn(async (key: string) => ({ key, cleared: true })),
-    };
-    const client = createGatewayApiClient({
-      config: DEFAULT_GATEWAY_CONFIG,
-      fetchFn,
-      nativeSessions,
-      nativeWebui,
-    });
-    await expect(client.sessions.patch("websocket:chat-1", { metadata: { pinned: true } })).resolves.toEqual({
-      key: "websocket:chat-1",
-      metadata: { pinned: true },
-      body: { metadata: { pinned: true } },
-    });
-    await expect(client.sessions.clear("websocket:chat-1")).resolves.toEqual({
-      key: "websocket:chat-1",
-      cleared: true,
-    });
-    await expect(client.sessions.delete("websocket:chat-1")).resolves.toEqual({
-      key: "websocket:chat-1",
-      deleted: true,
-    });
-    expect(nativeWebui.route).not.toHaveBeenCalled();
-    expect(fetchFn).not.toHaveBeenCalled();
-  });
-
-  test("prefers native WebUI session clear route when available", async () => {
-    const fetchFn = vi.fn(async () => new Response(JSON.stringify({ gateway: true }), { status: 200 }));
-    const nativeWebui = {
-      route: vi.fn(async (request: { method: string; path: string; body?: unknown }) => ({
-        key: "websocket:chat-1",
-        cleared: true,
-        request,
-      })),
-    };
-    const client = createGatewayApiClient({
-      config: DEFAULT_GATEWAY_CONFIG,
-      fetchFn,
-      nativeWebui,
-    });
-
-    await expect(client.sessions.clear("websocket:chat-1")).resolves.toMatchObject({
-      key: "websocket:chat-1",
-      cleared: true,
-    });
-    expect(nativeWebui.route).toHaveBeenCalledWith({
-      method: "POST",
-      path: "/api/sessions/websocket%3Achat-1/clear",
-    });
-    expect(fetchFn).not.toHaveBeenCalled();
-  });
-
-  test("prefers native WebUI session delete route when available", async () => {
-    const fetchFn = vi.fn(async () => new Response(JSON.stringify({ gateway: true }), { status: 200 }));
-    const nativeWebui = {
-      route: vi.fn(async (request: { method: string; path: string; body?: unknown }) => ({
-        key: "websocket:chat-1",
-        deleted: true,
-        request,
-      })),
-    };
-    const client = createGatewayApiClient({
-      config: DEFAULT_GATEWAY_CONFIG,
-      fetchFn,
-      nativeWebui,
-    });
-
-    await expect(client.sessions.delete("websocket:chat-1")).resolves.toMatchObject({
-      key: "websocket:chat-1",
-      deleted: true,
-    });
-    expect(nativeWebui.route).toHaveBeenCalledWith({
-      method: "DELETE",
-      path: "/api/sessions/websocket%3Achat-1",
-    });
-    expect(fetchFn).not.toHaveBeenCalled();
-  });
-
-  test("prefers native WebUI session profile route when available", async () => {
-    const fetchFn = vi.fn(async () => new Response(JSON.stringify({ gateway: true }), { status: 200 }));
-    const nativeWebui = {
-      route: vi.fn(async (request: { method: string; path: string; body?: unknown }) => ({
-        key: "websocket:chat-1",
-        profile: { display_name: "Ada" },
-        request,
-      })),
-    };
-    const client = createGatewayApiClient({
-      config: DEFAULT_GATEWAY_CONFIG,
-      fetchFn,
-      nativeWebui,
-    });
-
-    await expect(client.sessions.profile("websocket:chat-1")).resolves.toMatchObject({
-      key: "websocket:chat-1",
-      profile: { display_name: "Ada" },
-    });
-    expect(nativeWebui.route).toHaveBeenCalledWith({
-      method: "GET",
-      path: "/api/sessions/websocket%3Achat-1/profile",
-    });
-    expect(fetchFn).not.toHaveBeenCalled();
-  });
-
-  test("prefers native WebUI session patch route when available", async () => {
-    const fetchFn = vi.fn(async () => new Response(JSON.stringify({ gateway: true }), { status: 200 }));
-    const nativeWebui = {
-      route: vi.fn(async (request: { method: string; path: string; body?: unknown }) => ({
-        key: "websocket:chat-1",
-        metadata: { pinned: true },
-        updated_at: "2026-06-13T10:00:00.000Z",
-        request,
-      })),
-    };
-    const client = createGatewayApiClient({
-      config: DEFAULT_GATEWAY_CONFIG,
-      fetchFn,
-      nativeWebui,
-    });
-
-    await expect(client.sessions.patch("websocket:chat-1", { metadata: { pinned: true } })).resolves.toMatchObject({
-      key: "websocket:chat-1",
-      metadata: { pinned: true },
-      updated_at: "2026-06-13T10:00:00.000Z",
-    });
-    expect(nativeWebui.route).toHaveBeenCalledWith({
-      method: "PATCH",
-      path: "/api/sessions/websocket%3Achat-1",
-      body: { metadata: { pinned: true } },
-    });
-    expect(fetchFn).not.toHaveBeenCalled();
-  });
-
-  test("prefers native session branch adapter when available", async () => {
-    const fetchFn = vi.fn(async () => new Response(JSON.stringify({ gateway: true }), { status: 200 }));
-    const nativeSessions = {
-      list: vi.fn(),
-      messages: vi.fn(),
-      branch: vi.fn(async (body: unknown) => ({
-        key: "websocket:branch-1",
-        chat_id: "branch-1",
-        body,
-      })),
-    };
-    const client = createGatewayApiClient({
-      config: DEFAULT_GATEWAY_CONFIG,
-      fetchFn,
-      nativeSessions,
-    });
-
-    await expect(client.sessions.branch({ messages: [{ content: "Keep this" }] })).resolves.toMatchObject({
-      key: "websocket:branch-1",
-      chat_id: "branch-1",
-    });
-    expect(nativeSessions.branch).toHaveBeenCalledWith({ messages: [{ content: "Keep this" }] });
-    expect(fetchFn).not.toHaveBeenCalled();
-  });
-
-  test("posts session branch requests to the gateway when native routes are unavailable", async () => {
-    const fetchFn = vi.fn(async (url: RequestInfo | URL) => {
-      if (String(url).endsWith("/webui/bootstrap")) {
-        return new Response(JSON.stringify({ token: "token-1" }), { status: 200 });
-      }
-      return new Response(JSON.stringify({ key: "websocket:branch-1" }), { status: 200 });
-    });
-    const client = createGatewayApiClient({
-      config: DEFAULT_GATEWAY_CONFIG,
-      fetchFn,
-    });
-
-    await expect(client.sessions.branch({ messages: [{ content: "Keep this" }] })).resolves.toEqual({
-      key: "websocket:branch-1",
-    });
-    expect(fetchFn).toHaveBeenCalledWith(
-      "http://127.0.0.1:18790/api/sessions/branch",
-      expect.objectContaining({
-        method: "POST",
-      }),
-    );
-  });
-
-  test("prefers native Rust skills read operations when both native paths are available", async () => {
-    const fetchFn = vi.fn(async () => new Response(JSON.stringify({ gateway: true }), { status: 200 }));
-    const nativeWebui = {
-      route: vi.fn(async () => {
-        throw new Error("native WebUI skills route should not be used");
-      }),
-    };
-    const nativeSkills = {
-      list: vi.fn(async () => ({ skills: [{ name: "planner" }] })),
-      detail: vi.fn(async (name: string) => ({ name, content: "Plan." })),
-      create: vi.fn(async () => ({})),
-      update: vi.fn(async () => ({})),
-      delete: vi.fn(async () => ({})),
-      validate: vi.fn(async () => ({})),
-    };
-    const client = createGatewayApiClient({
-      config: DEFAULT_GATEWAY_CONFIG,
-      fetchFn,
-      nativeWebui,
-      nativeSkills,
-    });
-
-    await expect(client.skills.list()).resolves.toEqual({ skills: [{ name: "planner" }] });
-    await expect(client.skills.detail("planner/phase")).resolves.toEqual({ name: "planner/phase", content: "Plan." });
-    expect(nativeSkills.list).toHaveBeenCalledTimes(1);
-    expect(nativeSkills.detail).toHaveBeenCalledWith("planner/phase");
-    expect(nativeWebui.route).not.toHaveBeenCalled();
-    expect(fetchFn).not.toHaveBeenCalled();
-  });
-
-  test("prefers native Rust skills mutation operations when both native paths are available", async () => {
-    const fetchFn = vi.fn(async () => new Response(JSON.stringify({ gateway: true }), { status: 200 }));
-    const nativeWebui = {
-      route: vi.fn(async () => {
-        throw new Error("native WebUI skills route should not be used");
-      }),
-    };
-    const nativeSkills = {
-      list: vi.fn(async () => ({})),
-      detail: vi.fn(async () => ({})),
-      create: vi.fn(async (body: unknown) => ({ created: true, body })),
-      update: vi.fn(async (name: string, body: unknown) => ({ updated: true, name, body })),
-      delete: vi.fn(async (name: string) => ({ deleted: true, name })),
-      validate: vi.fn(async (name: string) => ({ name, valid: true })),
-    };
-    const client = createGatewayApiClient({
-      config: DEFAULT_GATEWAY_CONFIG,
-      fetchFn,
-      nativeWebui,
-      nativeSkills,
-    });
-
-    await expect(client.skills.create({ name: "planner", content: "Plan." })).resolves.toEqual({
-      created: true,
-      body: { name: "planner", content: "Plan." },
-    });
-    await expect(client.skills.update("planner/phase", { content: "Updated." })).resolves.toEqual({
-      updated: true,
-      name: "planner/phase",
-      body: { content: "Updated." },
-    });
-    await expect(client.skills.delete("planner/phase")).resolves.toEqual({
-      deleted: true,
-      name: "planner/phase",
-    });
-    await expect(client.skills.validate("planner/phase")).resolves.toEqual({
-      name: "planner/phase",
-      valid: true,
-    });
-    expect(nativeSkills.create).toHaveBeenCalledWith({ name: "planner", content: "Plan." });
-    expect(nativeSkills.update).toHaveBeenCalledWith("planner/phase", { content: "Updated." });
-    expect(nativeSkills.delete).toHaveBeenCalledWith("planner/phase");
-    expect(nativeSkills.validate).toHaveBeenCalledWith("planner/phase");
-    expect(nativeWebui.route).not.toHaveBeenCalled();
-    expect(fetchFn).not.toHaveBeenCalled();
-  });
-
-  test("prefers native WebUI Agent UI form routes when available", async () => {
-    const fetchFn = vi.fn(async () => new Response(JSON.stringify({ gateway: true }), { status: 200 }));
-    const nativeWebui = {
-      route: vi.fn(async (request: { method: string; path: string; body?: unknown }) => ({
-        native: true,
-        request,
-      })),
-    };
-    const client = createGatewayApiClient({
-      config: DEFAULT_GATEWAY_CONFIG,
-      fetchFn,
-      nativeWebui,
-    });
-
-    await expect(client.agentUi.submitForm("travel/plan", {
-      correlation: { session_key: "websocket:chat-1" },
-      values: { destination: "Paris" },
-    })).resolves.toEqual({
-      native: true,
-      request: {
-        method: "POST",
-        path: "/api/agent-ui/forms/travel%2Fplan/submit",
-        body: {
-          correlation: { session_key: "websocket:chat-1" },
-          values: { destination: "Paris" },
-        },
-      },
-    });
-    await expect(client.agentUi.cancelForm("travel/plan", {
-      correlation: { session_key: "websocket:chat-1" },
-    })).resolves.toEqual({
-      native: true,
-      request: {
-        method: "POST",
-        path: "/api/agent-ui/forms/travel%2Fplan/cancel",
-        body: { correlation: { session_key: "websocket:chat-1" } },
-      },
-    });
-    expect(fetchFn).not.toHaveBeenCalled();
-  });
-
-  test("prefers native Rust workspace file operations when both native paths are available", async () => {
-    const fetchFn = vi.fn(async () => new Response(JSON.stringify({ gateway: true }), { status: 200 }));
-    const nativeWebui = {
-      route: vi.fn(async () => {
-        throw new Error("native WebUI workspace route should not be used");
-      }),
-    };
-    const nativeWorkspace = {
-      files: vi.fn(async () => ({ items: [{ path: "docs/readme.md" }] })),
-      file: vi.fn(async (path: string) => ({ path, content: "# Readme\n" })),
-      putFile: vi.fn(async (path: string, body: unknown) => ({ path, saved: true, body })),
-      directory: vi.fn(async (request: { path: string }) => ({ result: { path: request.path, entries: [] } })),
-      fileChunk: vi.fn(async (request: { path: string }) => ({ result: { path: request.path, content_type: "text" } })),
-    };
-    const client = createGatewayApiClient({
-      config: DEFAULT_GATEWAY_CONFIG,
-      fetchFn,
-      nativeWebui,
-      nativeWorkspace,
-    });
-
-    await expect(client.workspace.files()).resolves.toEqual({
-      items: [{ path: "docs/readme.md" }],
-    });
-    await expect(client.workspace.file("docs/readme.md")).resolves.toEqual({
-      path: "docs/readme.md",
-      content: "# Readme\n",
-    });
-    await expect(client.workspace.putFile("docs/readme.md", {
-      content: "# Readme\n",
-      expected_updated_at: null,
-    })).resolves.toEqual({
-      path: "docs/readme.md",
-      saved: true,
-      body: {
-        content: "# Readme\n",
-        expected_updated_at: null,
-      },
-    });
-    await expect(client.workspace.directory({ path: "src" })).resolves.toEqual({
-      result: { path: "src", entries: [] },
-    });
-    await expect(client.workspace.fileChunk({ path: "src/main.ts" })).resolves.toEqual({
-      result: { path: "src/main.ts", content_type: "text" },
-    });
-    expect(nativeWorkspace.files).toHaveBeenCalledTimes(1);
-    expect(nativeWorkspace.file).toHaveBeenCalledWith("docs/readme.md");
-    expect(nativeWorkspace.putFile).toHaveBeenCalledWith("docs/readme.md", {
-      content: "# Readme\n",
-      expected_updated_at: null,
-    });
-    expect(nativeWorkspace.directory).toHaveBeenCalledWith({ path: "src" });
-    expect(nativeWorkspace.fileChunk).toHaveBeenCalledWith({ path: "src/main.ts" });
-    expect(nativeWebui.route).not.toHaveBeenCalled();
-    expect(fetchFn).not.toHaveBeenCalled();
-  });
-
-  test("prefers native WebUI OpenAI discovery routes when available", async () => {
-    const fetchFn = vi.fn(async () => new Response(JSON.stringify({ gateway: true }), { status: 200 }));
-    const nativeWebui = {
-      route: vi.fn(async (request: { method: string; path: string; body?: unknown }) => ({
-        native: true,
-        request,
-      })),
-    };
-    const client = createGatewayApiClient({
-      config: DEFAULT_GATEWAY_CONFIG,
-      fetchFn,
-      nativeWebui,
-    });
-    await expect(client.openAi.health()).resolves.toEqual({
-      native: true,
-      request: { method: "GET", path: "/health" },
-    });
-    await expect(client.openAi.models()).resolves.toEqual({
-      native: true,
-      request: { method: "GET", path: "/v1/models" },
-    });
-    expect(fetchFn).not.toHaveBeenCalled();
-  });
-
-  test("falls back to gateway skills operations when native skills are unavailable", async () => {
-    const fetchFn = vi.fn(async (url: RequestInfo | URL, _init?: RequestInit) => {
-      if (String(url).endsWith("/webui/bootstrap")) {
-        return new Response(JSON.stringify({ token: "token-1" }), { status: 200 });
-      }
-      return new Response(JSON.stringify({ gateway: true }), { status: 200 });
-    });
-    const client = createGatewayApiClient({
-      config: DEFAULT_GATEWAY_CONFIG,
-      fetchFn,
-      nativeSkills: {
-        list: async () => {
-          throw new Error("native unavailable");
-        },
-        detail: async () => {
-          throw new Error("native unavailable");
-        },
-        create: async () => {
-          throw new Error("native unavailable");
-        },
-        update: async () => {
-          throw new Error("native unavailable");
-        },
-        delete: async () => {
-          throw new Error("native unavailable");
-        },
-        validate: async () => {
-          throw new Error("native unavailable");
-        },
-      },
-    });
-
-    await expect(client.skills.list()).resolves.toEqual({ gateway: true });
-    await expect(client.skills.detail("planner/phase")).resolves.toEqual({ gateway: true });
-    await expect(client.skills.create({ name: "planner" })).resolves.toEqual({ gateway: true });
-    await expect(client.skills.update("planner/phase", { content: "# Updated" })).resolves.toEqual({ gateway: true });
-    await expect(client.skills.delete("planner/phase")).resolves.toEqual({ gateway: true });
-    await expect(client.skills.validate("planner/phase")).resolves.toEqual({ gateway: true });
-
-    expect(fetchFn.mock.calls.map((call) => String((call as unknown[])[0]))).toEqual([
-      "http://127.0.0.1:18790/webui/bootstrap",
-      "http://127.0.0.1:18790/api/skills",
-      "http://127.0.0.1:18790/api/skills/planner%2Fphase",
-      "http://127.0.0.1:18790/api/skills",
-      "http://127.0.0.1:18790/api/skills/planner%2Fphase",
-      "http://127.0.0.1:18790/api/skills/planner%2Fphase",
-      "http://127.0.0.1:18790/api/skills/planner%2Fphase/validate",
-    ]);
-  });
-
-  test("prefers native cowork route operations for migrated cowork paths", async () => {
-    const fetchFn = vi.fn(async (url: RequestInfo | URL, _init?: RequestInit) => {
-      if (String(url).endsWith("/webui/bootstrap")) {
-        return new Response(JSON.stringify({ token: "token-1" }), { status: 200 });
-      }
-      return new Response(JSON.stringify({ gateway: true }), { status: 200 });
-    });
-    const nativeCowork = {
-      route: vi.fn(async (request: { method: string; path: string; body?: unknown }) => ({
-        native: true,
-        request,
-      })),
-    };
-    const client = createGatewayApiClient({
-      config: DEFAULT_GATEWAY_CONFIG,
-      fetchFn,
-      nativeCowork,
-    });
-
-    await expect(client.cowork.sessions({ includeCompleted: true })).resolves.toMatchObject({ native: true });
-    await expect(client.cowork.session("cw_1")).resolves.toMatchObject({ native: true });
-    await expect(client.cowork.summary("cw_1")).resolves.toMatchObject({ native: true });
-    await expect(client.cowork.graph("cw_1")).resolves.toMatchObject({ native: true });
-    await expect(client.cowork.blueprint("cw_1")).resolves.toMatchObject({ native: true });
-    await expect(client.cowork.trace("cw_1")).resolves.toMatchObject({ native: true });
-    await expect(client.cowork.dag("cw_1")).resolves.toMatchObject({ native: true });
-    await expect(client.cowork.artifacts("cw_1")).resolves.toMatchObject({ native: true });
-    await expect(client.cowork.organization("cw_1")).resolves.toMatchObject({ native: true });
-    await expect(client.cowork.queues("cw_1")).resolves.toMatchObject({ native: true });
-    await expect(client.cowork.branches("cw_1")).resolves.toMatchObject({ native: true });
-    await expect(client.cowork.agentActivity("cw_1", "lead", { limit: 5 })).resolves.toMatchObject({ native: true });
-    await expect(client.cowork.observation("cw_1", "detail 1", { requesterAgentId: "reviewer" })).resolves.toMatchObject({ native: true });
-    await expect(client.cowork.create({ goal: "Native Cowork" })).resolves.toMatchObject({ native: true });
-    await expect(client.cowork.run("cw_1", { max_rounds: 4 })).resolves.toMatchObject({ native: true });
-    await expect(client.cowork.updateBudget("cw_1", { max_rounds: 4 })).resolves.toMatchObject({ native: true });
-    await expect(client.cowork.updateBudget("cw_1", { budgets: { max_tokens: 120 } }, { method: "PATCH" })).resolves.toMatchObject({ native: true });
-    await expect(client.cowork.action("cw_1", "pause")).resolves.toMatchObject({ native: true });
-    await expect(client.cowork.action("cw_1", "emergency-stop", { reason: "runaway delegation" })).resolves.toMatchObject({ native: true });
-    await expect(client.cowork.delete("cw_1")).resolves.toMatchObject({ native: true });
-    await expect(client.cowork.message("cw_1", { content: "Direct note", recipient_ids: ["lead"] })).resolves.toMatchObject({ native: true });
-    await expect(client.cowork.addTask("cw_1", { title: "Native task" })).resolves.toMatchObject({ native: true });
-    await expect(client.cowork.deriveBranch("cw_1", "branch 1", { target_architecture: "swarm" })).resolves.toMatchObject({ native: true });
-    await expect(client.cowork.taskAction("cw_1", "task/1", "assign", { assigned_agent_id: "lead" })).resolves.toMatchObject({ native: true });
-    await expect(client.cowork.workUnitAction("cw_1", "wu 1", "retry", { reason: "Retry" })).resolves.toMatchObject({ native: true });
-    await expect(client.cowork.selectBranch("cw_1", "branch 1")).resolves.toMatchObject({ native: true });
-    await expect(client.cowork.selectBranchResult("cw_1", "branch 1", { result_id: "result_1" })).resolves.toMatchObject({ native: true });
-    await expect(client.cowork.mergeBranchResults("cw_1", { branch_ids: ["branch 1", "branch 2"] })).resolves.toMatchObject({ native: true });
-    await expect(client.cowork.selectFinalResult("cw_1", { branch_id: "branch 1", result_id: "result_1" })).resolves.toMatchObject({ native: true });
-    await expect(client.cowork.mergeFinalResult("cw_1", { branch_ids: ["branch 1", "branch 2"] })).resolves.toMatchObject({ native: true });
-
-    expect(nativeCowork.route).toHaveBeenCalledWith({
-      method: "GET",
-      path: "/api/cowork/sessions",
-      query: { include_completed: "true" },
-    });
-    expect(nativeCowork.route).toHaveBeenCalledWith({
-      method: "GET",
-      path: "/api/cowork/sessions/cw_1",
-    });
-    expect(nativeCowork.route).toHaveBeenCalledWith({
-      method: "GET",
-      path: "/api/cowork/sessions/cw_1/summary",
-    });
-    expect(nativeCowork.route).toHaveBeenCalledWith({
-      method: "GET",
-      path: "/api/cowork/sessions/cw_1/graph",
-    });
-    expect(nativeCowork.route).toHaveBeenCalledWith({
-      method: "GET",
-      path: "/api/cowork/sessions/cw_1/blueprint",
-    });
-    expect(nativeCowork.route).toHaveBeenCalledWith({
-      method: "GET",
-      path: "/api/cowork/sessions/cw_1/trace",
-    });
-    expect(nativeCowork.route).toHaveBeenCalledWith({
-      method: "GET",
-      path: "/api/cowork/sessions/cw_1/dag",
-    });
-    expect(nativeCowork.route).toHaveBeenCalledWith({
-      method: "GET",
-      path: "/api/cowork/sessions/cw_1/artifacts",
-    });
-    expect(nativeCowork.route).toHaveBeenCalledWith({
-      method: "GET",
-      path: "/api/cowork/sessions/cw_1/organization",
-    });
-    expect(nativeCowork.route).toHaveBeenCalledWith({
-      method: "GET",
-      path: "/api/cowork/sessions/cw_1/queues",
-    });
-    expect(nativeCowork.route).toHaveBeenCalledWith({
-      method: "GET",
-      path: "/api/cowork/sessions/cw_1/branches",
-    });
-    expect(nativeCowork.route).toHaveBeenCalledWith({
-      method: "GET",
-      path: "/api/cowork/sessions/cw_1/agents/lead/activity",
-      query: { limit: "5" },
-    });
-    expect(nativeCowork.route).toHaveBeenCalledWith({
-      method: "GET",
-      path: "/api/cowork/sessions/cw_1/observations/detail%201",
-      query: { agent_id: "reviewer" },
-    });
-    expect(nativeCowork.route).toHaveBeenCalledWith({
-      method: "POST",
-      path: "/api/cowork/sessions",
-      body: { goal: "Native Cowork" },
-    });
-    expect(nativeCowork.route).toHaveBeenCalledWith({
-      method: "POST",
-      path: "/api/cowork/sessions/cw_1/run",
-      body: { max_rounds: 4 },
-    });
-    expect(nativeCowork.route).toHaveBeenCalledWith({
-      method: "POST",
-      path: "/api/cowork/sessions/cw_1/budget",
-      body: { max_rounds: 4 },
-    });
-    expect(nativeCowork.route).toHaveBeenCalledWith({
-      method: "PATCH",
-      path: "/api/cowork/sessions/cw_1/budget",
-      body: { budgets: { max_tokens: 120 } },
-    });
-    expect(nativeCowork.route).toHaveBeenCalledWith({
-      method: "POST",
-      path: "/api/cowork/sessions/cw_1/pause",
-    });
-    expect(nativeCowork.route).toHaveBeenCalledWith({
-      method: "POST",
-      path: "/api/cowork/sessions/cw_1/emergency-stop",
-      body: { reason: "runaway delegation" },
-    });
-    expect(nativeCowork.route).toHaveBeenCalledWith({
-      method: "DELETE",
-      path: "/api/cowork/sessions/cw_1",
-    });
-    expect(nativeCowork.route).toHaveBeenCalledWith({
-      method: "POST",
-      path: "/api/cowork/sessions/cw_1/messages",
-      body: { content: "Direct note", recipient_ids: ["lead"] },
-    });
-    expect(nativeCowork.route).toHaveBeenCalledWith({
-      method: "POST",
-      path: "/api/cowork/sessions/cw_1/tasks",
-      body: { title: "Native task" },
-    });
-    expect(nativeCowork.route).toHaveBeenCalledWith({
-      method: "POST",
-      path: "/api/cowork/sessions/cw_1/branches/branch%201/derive",
-      body: { target_architecture: "swarm" },
-    });
-    expect(nativeCowork.route).toHaveBeenCalledWith({
-      method: "POST",
-      path: "/api/cowork/sessions/cw_1/tasks/task%2F1/assign",
-      body: { assigned_agent_id: "lead" },
-    });
-    expect(nativeCowork.route).toHaveBeenCalledWith({
-      method: "POST",
-      path: "/api/cowork/sessions/cw_1/work-units/wu%201/retry",
-      body: { reason: "Retry" },
-    });
-    expect(nativeCowork.route).toHaveBeenCalledWith({
-      method: "POST",
-      path: "/api/cowork/sessions/cw_1/branches/branch%201/select",
-    });
-    expect(nativeCowork.route).toHaveBeenCalledWith({
-      method: "POST",
-      path: "/api/cowork/sessions/cw_1/branches/branch%201/result/select-final",
-      body: { result_id: "result_1" },
-    });
-    expect(nativeCowork.route).toHaveBeenCalledWith({
-      method: "POST",
-      path: "/api/cowork/sessions/cw_1/branch-results/merge",
-      body: { branch_ids: ["branch 1", "branch 2"] },
-    });
-    expect(nativeCowork.route).toHaveBeenCalledWith({
-      method: "POST",
-      path: "/api/cowork/sessions/cw_1/final-result/select",
-      body: { branch_id: "branch 1", result_id: "result_1" },
-    });
-    expect(nativeCowork.route).toHaveBeenCalledWith({
-      method: "POST",
-      path: "/api/cowork/sessions/cw_1/final-result/merge",
-      body: { branch_ids: ["branch 1", "branch 2"] },
-    });
-    expect(fetchFn).not.toHaveBeenCalled();
-  });
-
-  test("passes cowork route query parameters through the native route query field", async () => {
-    const fetchFn = vi.fn(async (url: RequestInfo | URL, _init?: RequestInit) => {
-      if (String(url).endsWith("/webui/bootstrap")) {
-        return new Response(JSON.stringify({ token: "token-1" }), { status: 200 });
-      }
-      return new Response(JSON.stringify({ gateway: true }), { status: 200 });
-    });
-    const nativeCowork = {
-      route: vi.fn(async (request: { method: string; path: string; query?: Record<string, unknown> }) => ({
-        native: true,
-        request,
-      })),
-    };
-    const client = createGatewayApiClient({
-      config: DEFAULT_GATEWAY_CONFIG,
-      fetchFn,
-      nativeCowork,
-    });
-
-    await expect(client.cowork.sessions({ includeCompleted: true, originChatId: "chat/1" })).resolves.toMatchObject({
-      native: true,
-    });
-    await expect(client.cowork.agentActivity("cw_1", "lead", { limit: 5 })).resolves.toMatchObject({
-      native: true,
-    });
-
-    expect(nativeCowork.route).toHaveBeenCalledWith({
-      method: "GET",
-      path: "/api/cowork/sessions",
-      query: { include_completed: "true", origin_chat_id: "chat/1" },
-    });
-    expect(nativeCowork.route).toHaveBeenCalledWith({
-      method: "GET",
-      path: "/api/cowork/sessions/cw_1/agents/lead/activity",
-      query: { limit: "5" },
-    });
-    expect(fetchFn).not.toHaveBeenCalled();
-  });
-
-  test("does not fall back to HTTP when native cowork run fails", async () => {
-    const fetchFn = vi.fn(async () => new Response(JSON.stringify({ gateway: true }), { status: 200 }));
-    const nativeCowork = {
-      route: vi.fn(async () => {
-        throw new Error("native unavailable");
-      }),
-    };
-    const client = createGatewayApiClient({
-      config: DEFAULT_GATEWAY_CONFIG,
-      fetchFn,
-      nativeCowork,
-    });
-
-    await expect(client.cowork.run("cw_1", { max_rounds: 4 })).rejects.toThrow("native unavailable");
-
-    expect(nativeCowork.route).toHaveBeenCalledWith({
-      method: "POST",
-      path: "/api/cowork/sessions/cw_1/run",
-      body: { max_rounds: 4 },
-    });
-    expect(fetchFn).not.toHaveBeenCalled();
-  });
-
-  test("refreshes the gateway token before authenticated requests when the session is near expiry", async () => {
-    vi.useFakeTimers();
-    vi.setSystemTime(new Date("2026-06-08T10:00:00.000Z"));
-    try {
-      const fetchFn = vi.fn(async (url: RequestInfo | URL, init?: RequestInit) => {
-        const path = new URL(String(url)).pathname;
-        if (path === "/webui/bootstrap") {
-          return new Response(
-            JSON.stringify({
-              token: "token-1",
-              refresh_token_path: "/webui/refresh-token",
-              token_ttl_s: 300,
-            }),
-            { status: 200 },
-          );
-        }
-        if (path === "/webui/refresh-token") {
-          expect(init).toMatchObject({
-            method: "POST",
-            headers: expect.objectContaining({ Authorization: "Bearer token-1" }),
-          });
-          return new Response(
-            JSON.stringify({
-              token: "token-1",
-              refresh_token_path: "/webui/refresh-token",
-              token_ttl_s: 300,
-            }),
-            { status: 200 },
-          );
-        }
-        return new Response(JSON.stringify({ items: [] }), { status: 200 });
-      });
-      const client = createGatewayApiClient({
-        config: DEFAULT_GATEWAY_CONFIG,
-        fetchFn,
-      });
-
-      await client.sessions.list();
-      vi.setSystemTime(new Date("2026-06-08T10:04:15.000Z"));
-      await client.sessions.list();
-
-      expect(fetchFn.mock.calls.map((call) => String(call[0]))).toEqual([
-        "http://127.0.0.1:18790/webui/bootstrap",
-        "http://127.0.0.1:18790/api/sessions",
-        "http://127.0.0.1:18790/webui/refresh-token",
-        "http://127.0.0.1:18790/api/sessions",
-      ]);
-    } finally {
-      vi.useRealTimers();
-    }
-  });
-
-  test("prefers native WebUI refresh-token route when the session is near expiry", async () => {
-    vi.useFakeTimers();
-    vi.setSystemTime(new Date("2026-06-08T10:00:00.000Z"));
-    try {
-      const fetchFn = vi.fn(async (url: RequestInfo | URL) => {
-        const path = new URL(String(url)).pathname;
-        if (path === "/webui/bootstrap") {
-          return new Response(
-            JSON.stringify({
-              token: "token-1",
-              refresh_token_path: "/webui/refresh-token",
-              token_ttl_s: 300,
-            }),
-            { status: 200 },
-          );
-        }
-        if (path === "/api/sessions") {
-          return new Response(JSON.stringify({ items: [] }), { status: 200 });
-        }
-        return new Response(JSON.stringify({ error: "unexpected route" }), { status: 404 });
-      });
-      const nativeWebui = {
-        route: vi.fn(async (request: { method: string; path: string; headers?: Record<string, unknown> }) => {
-          if (request.path === "/webui/refresh-token") {
-            return {
-              token: "token-2",
-              refresh_token_path: "/webui/refresh-token",
-              token_ttl_s: 300,
-              request,
-            };
-          }
-          throw new Error("native sessions unavailable");
-        }),
-      };
-      const client = createGatewayApiClient({
-        config: DEFAULT_GATEWAY_CONFIG,
-        fetchFn,
-        nativeWebui,
-      });
-
-      await client.sessions.list();
-      vi.setSystemTime(new Date("2026-06-08T10:04:15.000Z"));
-      await client.sessions.list();
-
-      expect(nativeWebui.route).toHaveBeenCalledWith({
-        method: "POST",
-        path: "/webui/refresh-token",
-        headers: { Authorization: "Bearer token-1" },
-      });
-      expect(fetchFn.mock.calls.map((call) => String(call[0]))).toEqual([
-        "http://127.0.0.1:18790/webui/bootstrap",
-        "http://127.0.0.1:18790/api/sessions",
-        "http://127.0.0.1:18790/api/sessions",
-      ]);
-    } finally {
-      vi.useRealTimers();
-    }
-  });
-
-  test("retries gateway bootstrap after an earlier bootstrap failure", async () => {
-    let bootstrapAttempts = 0;
-    const fetchFn = vi.fn(async (url: RequestInfo | URL, _init?: RequestInit) => {
-      const path = new URL(String(url)).pathname;
-      if (path === "/webui/bootstrap") {
-        bootstrapAttempts += 1;
-        if (bootstrapAttempts === 1) {
-          throw new TypeError("Failed to fetch");
-        }
-        return new Response(JSON.stringify({ token: "token-2", token_ttl_s: 300 }), { status: 200 });
-      }
-      if (path === "/api/approvals") {
-        return new Response(JSON.stringify({ approvals: [] }), { status: 200 });
-      }
-      if (path === "/api/sessions/WebSocket%3Achat-live/messages") {
-        return new Response(JSON.stringify({ messages: [] }), { status: 200 });
-      }
-      return new Response(JSON.stringify({ error: "unexpected route" }), { status: 404 });
-    });
-    const client = createGatewayApiClient({
-      config: DEFAULT_GATEWAY_CONFIG,
-      fetchFn,
-    });
-
-    await expect(client.tools.approvals()).rejects.toThrow("Gateway bootstrap failed: Failed to fetch");
-    await expect(client.sessions.messages("WebSocket:chat-live")).resolves.toEqual({ messages: [] });
-
-    expect(fetchFn.mock.calls.map((call) => String(call[0]))).toEqual([
-      "http://127.0.0.1:18790/webui/bootstrap",
-      "http://127.0.0.1:18790/webui/bootstrap",
-      "http://127.0.0.1:18790/api/sessions/WebSocket%3Achat-live/messages",
-    ]);
-    expect(fetchFn.mock.calls[2][1]).toMatchObject({
-      headers: expect.objectContaining({ Authorization: "Bearer token-2" }),
-    });
-  });
-
-  test("reports a clear gateway bootstrap timeout reason", async () => {
-    const fetchFn = vi.fn(async (_url: RequestInfo | URL, init?: RequestInit) => {
-      await new Promise((_resolve, reject) => {
-        const signal = init?.signal;
-        signal?.addEventListener("abort", () => reject(signal.reason), { once: true });
-      });
-      return new Response(JSON.stringify({ token: "late-token" }), { status: 200 });
-    });
-    const client = createGatewayApiClient({
-      config: resolveGatewayConfig({
-        ...DEFAULT_GATEWAY_CONFIG,
-        requestTimeoutMs: 5,
-      }),
-      fetchFn,
-    });
-
-    await expect(client.sessions.list()).rejects.toThrow("Gateway bootstrap timed out after 5 ms");
-  });
-
-  test("bootstraps a fresh token and retries once when an authenticated request receives 401", async () => {
-    const fetchFn = vi.fn(async (url: RequestInfo | URL, init?: RequestInit) => {
-      const path = new URL(String(url)).pathname;
-      if (path === "/webui/bootstrap") {
-        const token = fetchFn.mock.calls.filter((call) => String(call[0]).endsWith("/webui/bootstrap")).length;
-        return new Response(JSON.stringify({ token: `token-${token}`, token_ttl_s: 300 }), { status: 200 });
-      }
-      if (path === "/api/sessions") {
-        const authorization = (init?.headers as Record<string, string> | undefined)?.Authorization;
-        if (authorization === "Bearer token-1") {
-          return new Response(JSON.stringify({ error: "unauthorized" }), { status: 401 });
-        }
-        return new Response(JSON.stringify({ items: [] }), { status: 200 });
-      }
-      return new Response(JSON.stringify({ ok: true }), { status: 200 });
-    });
-    const client = createGatewayApiClient({
-      config: DEFAULT_GATEWAY_CONFIG,
-      fetchFn,
-    });
-
-    await expect(client.sessions.list()).resolves.toEqual({ items: [] });
-
-    expect(fetchFn.mock.calls.map((call) => String(call[0]))).toEqual([
-      "http://127.0.0.1:18790/webui/bootstrap",
-      "http://127.0.0.1:18790/api/sessions",
-      "http://127.0.0.1:18790/webui/bootstrap",
-      "http://127.0.0.1:18790/api/sessions",
-    ]);
-    expect(fetchFn.mock.calls[1][1]).toMatchObject({
-      headers: expect.objectContaining({ Authorization: "Bearer token-1" }),
-    });
-    expect(fetchFn.mock.calls[3][1]).toMatchObject({
-      headers: expect.objectContaining({ Authorization: "Bearer token-2" }),
-    });
-  });
-});
-
-describe("gateway WebSocket client", () => {
-  test("creates outbound chat control messages", () => {
-    expect(createGatewaySocketMessage.newChat()).toEqual({ type: "new_chat" });
-    expect(createGatewaySocketMessage.attach("chat-1")).toEqual({
-      type: "attach",
-      chat_id: "chat-1",
-    });
-    expect(createGatewaySocketMessage.message("chat-1", "hello", "deepseek-reasoner")).toEqual({
-      type: "message",
-      chat_id: "chat-1",
-      content: "hello",
-      model: "deepseek-reasoner",
-    });
-    expect(createGatewaySocketMessage.message("chat-1", "explain", undefined, "client-1", [{
-      detail: "TinyOS file selection",
-      evidenceId: "item-1",
-      kind: "reference",
-      sourceLine: 3,
-      sourcePath: "src/main.ts",
-      sourceText: "const value = 1;",
-      title: "src/main.ts · L3",
-      type: "tinyos.file",
-    }])).toEqual({
-      type: "message",
-      chat_id: "chat-1",
-      client_event_id: "client-1",
-      content: "explain",
-      references: [expect.objectContaining({ evidenceId: "item-1", type: "tinyos.file" })],
-    });
-    expect(createGatewaySocketMessage.interrupt("chat-1", {
-      schemaVersion: "tinybot.command.v1",
-      commandId: "command-1",
-      issuedAt: "2026-07-13T00:00:00Z",
-      kind: "agent.cancel",
-      source: { control: "stop-response", surface: "chat" },
-      target: {
-        sessionId: "websocket:chat-1",
-        threadId: "thread-1",
-        turnId: "turn-1",
-      },
-    })).toEqual({
-      type: "interrupt",
-      chat_id: "chat-1",
-      command_id: "command-1",
-      command_kind: "agent.cancel",
-      session_id: "websocket:chat-1",
-      source: { control: "stop-response", surface: "chat" },
-      thread_id: "thread-1",
-      turn_id: "turn-1",
-    });
-    expect(createGatewaySocketMessage.command("chat-1", {
-      schemaVersion: "tinybot.command.v1",
-      commandId: "command-pause-1",
-      issuedAt: "2026-07-13T00:00:00Z",
-      kind: "agent.pause",
-      source: { control: "system-bar-pause", surface: "tinyos" },
-      target: { turnId: "turn-1", sessionId: "websocket:chat-1" },
-    })).toEqual({
-      type: "command",
-      chat_id: "chat-1",
-      command_id: "command-pause-1",
-      command_kind: "agent.pause",
-      session_id: "websocket:chat-1",
-      source: { control: "system-bar-pause", surface: "tinyos" },
-      turn_id: "turn-1",
-    });
-    expect(createGatewaySocketMessage.command("chat-1", {
-      schemaVersion: "tinybot.command.v1",
-      commandId: "command-approval-1",
-      issuedAt: "2026-07-13T00:00:00Z",
-      kind: "approval.resolve",
-      source: { control: "inspector-approval", surface: "tinyos" },
-      target: { turnId: "turn-1", sessionId: "websocket:chat-1" },
-      approval: { approvalId: "approval-1", approved: true, scope: "session" },
-    })).toEqual({
-      type: "command",
-      chat_id: "chat-1",
-      command_id: "command-approval-1",
-      command_kind: "approval.resolve",
-      turn_id: "turn-1",
-      session_id: "websocket:chat-1",
-      source: { control: "inspector-approval", surface: "tinyos" },
-      approval_id: "approval-1",
-      approved: true,
-      scope: "session",
-    });
-    expect(createGatewaySocketMessage.command("chat-1", {
-      schemaVersion: "tinybot.command.v1",
-      commandId: "command-form-1",
-      issuedAt: "2026-07-13T00:00:00Z",
-      kind: "form.submit",
-      source: { control: "system-form", surface: "tinyos" },
-      target: { turnId: "turn-1", sessionId: "websocket:chat-1" },
-      form: { formId: "travel-preferences-1", values: { destination: "Singapore" } },
-    })).toEqual({
-      type: "command",
-      chat_id: "chat-1",
-      command_id: "command-form-1",
-      command_kind: "form.submit",
-      session_id: "websocket:chat-1",
-      turn_id: "turn-1",
-      source: { control: "system-form", surface: "tinyos" },
-      form_id: "travel-preferences-1",
-      values: { destination: "Singapore" },
-    });
-    expect(createGatewaySocketMessage.command("chat-1", {
-      schemaVersion: "tinybot.command.v1",
-      commandId: "command-form-cancel-1",
-      issuedAt: "2026-07-13T00:00:00Z",
-      kind: "form.cancel",
-      source: { control: "chat-form", surface: "chat" },
-      target: { turnId: "turn-1", sessionId: "websocket:chat-1" },
-      form: { formId: "travel-preferences-1" },
-    })).toEqual({
-      type: "command",
-      chat_id: "chat-1",
-      command_id: "command-form-cancel-1",
-      command_kind: "form.cancel",
-      session_id: "websocket:chat-1",
-      turn_id: "turn-1",
-      source: { control: "chat-form", surface: "chat" },
-      form_id: "travel-preferences-1",
-    });
-    expect(createGatewaySocketMessage.command("chat-1", {
-      schemaVersion: "tinybot.command.v1",
-      commandId: "command-retry-1",
-      issuedAt: "2026-07-13T00:00:00Z",
-      kind: "operation.retry",
-      source: { control: "operation-shelf", surface: "tinyos" },
-      target: { turnId: "turn-retry-1", sessionId: "websocket:chat-1" },
-      operation: { itemId: "turn-failed:error", turnId: "turn-failed" },
-    })).toEqual({
-      type: "command",
-      chat_id: "chat-1",
-      command_id: "command-retry-1",
-      command_kind: "operation.retry",
-      turn_id: "turn-retry-1",
-      session_id: "websocket:chat-1",
-      source: { control: "operation-shelf", surface: "tinyos" },
-      source_turn_id: "turn-failed",
-      item_id: "turn-failed:error",
-    });
-    expect(createGatewaySocketMessage.command("chat-1", {
-      schemaVersion: "tinybot.command.v1",
-      commandId: "command-request-1",
-      issuedAt: "2026-07-13T00:00:00Z",
-      kind: "agent.request_change",
-      source: { control: "files-explain-selection", surface: "tinyos" },
-      target: { turnId: "turn-request-1", sessionId: "websocket:chat-1" },
-      request: {
-        instruction: "Explain this selection.",
-        observedTurnId: "turn-completed-1",
-        references: [{
-          detail: "TinyOS file selection",
-          kind: "reference",
-          sourceEndLine: 3,
-          sourceLine: 2,
-          sourcePath: "src/main.ts",
-          sourceText: "return value;",
-          title: "src/main.ts · L2–3",
-          type: "tinyos.file",
-        }],
-      },
-    })).toEqual({
-      type: "command",
-      chat_id: "chat-1",
-      command_id: "command-request-1",
-      command_kind: "agent.request_change",
-      turn_id: "turn-request-1",
-      session_id: "websocket:chat-1",
-      source: { control: "files-explain-selection", surface: "tinyos" },
-      instruction: "Explain this selection.",
-      observed_turn_id: "turn-completed-1",
-      references: [{
-        detail: "TinyOS file selection",
-        kind: "reference",
-        sourceEndLine: 3,
-        sourceLine: 2,
-        sourcePath: "src/main.ts",
-        sourceText: "return value;",
-        title: "src/main.ts · L2–3",
-        type: "tinyos.file",
-      }],
-    });
-    expect(createGatewaySocketMessage.command("chat-1", {
-      schemaVersion: "tinybot.command.v1",
-      commandId: "command-file-save-1",
-      issuedAt: "2026-07-13T00:00:00Z",
-      kind: "file.save",
-      source: { control: "files-editor", surface: "tinyos" },
-      target: { operationId: "tinyos-host-file-1", sessionId: "websocket:chat-1" },
-      file: {
-        baseRevision: "metadata:12:34",
-        confirmed: true,
-        content: "updated\n",
-        createOnly: false,
-        path: "notes/today.md",
-      },
-    })).toEqual({
-      type: "command",
-      chat_id: "chat-1",
-      command_id: "command-file-save-1",
-      command_kind: "file.save",
-      operation_id: "tinyos-host-file-1",
-      session_id: "websocket:chat-1",
-      source: { control: "files-editor", surface: "tinyos" },
-      path: "notes/today.md",
-      content: "updated\n",
-      create_only: false,
-      confirmed: true,
-      base_revision: "metadata:12:34",
-    });
-    expect(createGatewaySocketMessage.command("chat-1", {
-      schemaVersion: "tinybot.command.v1",
-      commandId: "command-terminal-1",
-      issuedAt: "2026-07-13T00:00:00Z",
-      kind: "terminal.execute",
-      source: { control: "terminal-command", surface: "tinyos" },
-      target: { operationId: "tinyos-host-terminal-1", sessionId: "websocket:chat-1" },
-      terminal: { command: "npm test", confirmed: true, cwd: "apps/desktop" },
-    })).toEqual({
-      type: "command",
-      chat_id: "chat-1",
-      command_id: "command-terminal-1",
-      command_kind: "terminal.execute",
-      operation_id: "tinyos-host-terminal-1",
-      session_id: "websocket:chat-1",
-      source: { control: "terminal-command", surface: "tinyos" },
-      command: "npm test",
-      confirmed: true,
-      cwd: "apps/desktop",
-    });
-  });
-
-  test("queues outbound messages until the socket is open", () => {
-    const sent: string[] = [];
-    const queue: unknown[] = [];
-    const connectingSocket = {
-      readyState: 0,
-      send: (value: string) => sent.push(value),
-    };
-    const openSocket = {
-      readyState: 1,
-      send: (value: string) => sent.push(value),
-    };
-
-    expect(sendGatewaySocketJson(connectingSocket, { type: "new_chat" }, queue)).toBe("queued");
-    expect(sent).toEqual([]);
-    expect(queue).toEqual([{ type: "new_chat" }]);
-
-    expect(flushGatewaySocketQueue(openSocket, queue)).toBe(1);
-    expect(sent).toEqual([JSON.stringify({ type: "new_chat" })]);
-    expect(queue).toEqual([]);
-  });
-
-  test("normalizes browser, agent-ui, and agent event frames while ignoring legacy message streams", () => {
-    expect(normalizeGatewayFrame({ event: "attached", chat_id: "chat-1" })).toMatchObject({
-      kind: "attached",
-      chatId: "chat-1",
-    });
-    expect(normalizeGatewayFrame({ event: "command_accepted", chat_id: "chat-1", command_id: "command-1" })).toMatchObject({
-      kind: "command.accepted",
-      chatId: "chat-1",
-      commandId: "command-1",
-    });
-    expect(normalizeGatewayFrame({ event: "command_canonical_updated", chat_id: "chat-1", command_id: "command-1" })).toMatchObject({
-      kind: "command.canonical-updated",
-      chatId: "chat-1",
-      commandId: "command-1",
-    });
-    expect(normalizeGatewayFrame({ event: "error", command_id: "command-1", message: "not active" })).toMatchObject({
-      kind: "error",
-      commandId: "command-1",
-      message: "not active",
-    });
-    expect(normalizeGatewayFrame({ event: "delta", text: "hi", message_id: "m1" })).toMatchObject({
-      kind: "unknown",
-      event: "delta",
-    });
-    expect(normalizeGatewayFrame({ event: "delta", text: "plan", is_reasoning: true })).toMatchObject({
-      kind: "unknown",
-      event: "delta",
-    });
-    expect(normalizeGatewayFrame({ event: "message", text: "done", message_id: "m2" })).toMatchObject({
-      kind: "unknown",
-      event: "message",
-    });
-    expect(normalizeGatewayFrame({ event: "stream_end", chat_id: "chat-1" })).toMatchObject({
-      kind: "unknown",
-      event: "stream_end",
-    });
-    expect(normalizeGatewayFrame({ event: "usage", chat_id: "chat-1", usage: { total_tokens: 16384 } })).toMatchObject({
-      kind: "usage",
-      chatId: "chat-1",
-      tokenUsage: "16384 tokens",
-    });
-    expect(
-      normalizeGatewayFrame({
-        event: "usage",
-        chat_id: "chat-1",
-        usage: { total_tokens: 16384, context_window_tokens: 65536 },
-      }),
-    ).toMatchObject({
-      kind: "usage",
-      chatId: "chat-1",
-      tokenUsage: "16384 / 65536 tokens (25%)",
-    });
-    expect(normalizeGatewayFrame({ event: "browser_frame", image: "data:image/png;base64,x" })).toMatchObject({
-      kind: "browser.frame",
-    });
-    expect(normalizeGatewayFrame({ event: "agent_ui_form", form: { form_id: "form-1" } })).toMatchObject({
-      kind: "agent-ui.form",
-    });
-    expect(
-      normalizeGatewayFrame({
-        event: "agent_event",
-        schema_version: "tinybot.agent_event.v1",
-        event_id: "event-tool",
-        event_type: "tool.call.started",
-        chat_id: "chat-1",
-        turn_id: "turn-1",
-        payload: { name: "read_file" },
-      }),
-    ).toMatchObject({
-      kind: "agent.event",
-      chatId: "chat-1",
-    });
-    expect(
-      normalizeGatewayFrame({
-        event: "agent_ui_event",
-        agent_ui_event: { event_type: "ui.form.requested", payload: { form_id: "form-1" } },
-      }),
-    ).toMatchObject({
-      kind: "agent-ui.event",
-      eventType: "ui.form.requested",
-    });
-    expect(
-      normalizeGatewayFrame({
-        event: "agent_ui_event",
-        chat_id: "chat-1",
-        agent_ui_event: {
-          event_type: "message.delta",
-          chat_id: "chat-1",
-          message_id: "m3",
-          payload: { text: "streamed" },
-        },
-      }),
-    ).toMatchObject({
-      kind: "agent-ui.event",
-      eventType: "message.delta",
-    });
-    expect(
-      normalizeGatewayFrame({
-        event: "agent_ui_event",
-        chat_id: "chat-1",
-        agent_ui_event: {
-          event_type: "usage.updated",
-          payload: { usage: { total_tokens: 32768, context_window_tokens: 65536 } },
-        },
-      }),
-    ).toMatchObject({
-      kind: "usage",
-      chatId: "chat-1",
-      tokenUsage: "32768 / 65536 tokens (50%)",
-      usage: { total_tokens: 32768, context_window_tokens: 65536 },
-    });
-
-    expect(
-      normalizeGatewayFrame({
-        event: "cowork_stream",
-        chat_id: "chat-1",
-        session_id: "session-1",
-        agent_id: "agent-1",
-        step_id: "step-1",
-        phase: "delta",
-        status: "running",
-        sequence: 3,
-        text: "live answer",
-        completed: false,
-      }),
-    ).toMatchObject({
-      kind: "unknown",
-      event: "cowork_stream",
-    });
-
-    expect(
-      normalizeGatewayFrame({
-        event: "cowork_mailbox_stream",
-        chat_id: "chat-1",
-        session_id: "session-1",
-        draft_id: "draft-1",
-        tool_call_id: "call-1",
-        phase: "terminal",
-        status: "completed",
-        completed: true,
-      }),
-    ).toMatchObject({
-      kind: "unknown",
-      event: "cowork_mailbox_stream",
-    });
-  });
-});
