--- src/app-core/gateway/gatewayHttpClient.ts
@@ -1,1243 +1,0 @@
-import { DEFAULT_GATEWAY_CONFIG, type GatewayConfig } from "./gatewayConfig";
-import { logDesktopNativeDebug } from "../native/desktopNativeChatDebug";
-
-type FetchFn = typeof fetch;
-
-type ClientOptions = {
-  config?: GatewayConfig;
-  fetchFn?: FetchFn;
-  nativeConfig?: NativeConfigApi;
-  nativeSessions?: NativeSessionsApi;
-  nativeThreads?: NativeThreadsApi;
-  nativeSkills?: NativeSkillsApi;
-  nativeWorkspace?: NativeWorkspaceApi;
-  nativeCowork?: NativeCoworkApi;
-  nativeWebui?: NativeWebuiApi;
-  tsCoworkRuntime?: TsCoworkRuntimeRollout;
-};
-
-export type TsCoworkRuntimeRollout = {
-  enabled?: boolean;
-  readOnlySnapshot?: boolean;
-  mutations?: boolean;
-  scheduler?: boolean;
-  swarm?: boolean;
-};
-
-export const DEFAULT_TS_COWORK_RUNTIME_ROLLOUT: Required<TsCoworkRuntimeRollout> = {
-  enabled: true,
-  readOnlySnapshot: true,
-  mutations: true,
-  scheduler: true,
-  swarm: true,
-};
-
-export function resolveTsCoworkRuntimeRollout(config: unknown): Required<TsCoworkRuntimeRollout> {
-  const desktop = asRecord(config)?.desktop;
-  const rollout = asRecord(asRecord(desktop)?.tsCoworkRuntime ?? asRecord(desktop)?.ts_cowork_runtime);
-  return {
-    enabled: booleanValue(rollout?.enabled, DEFAULT_TS_COWORK_RUNTIME_ROLLOUT.enabled),
-    readOnlySnapshot: booleanValue(
-      rollout?.readOnlySnapshot ?? rollout?.read_only_snapshot,
-      DEFAULT_TS_COWORK_RUNTIME_ROLLOUT.readOnlySnapshot,
-    ),
-    mutations: booleanValue(rollout?.mutations, DEFAULT_TS_COWORK_RUNTIME_ROLLOUT.mutations),
-    scheduler: booleanValue(rollout?.scheduler, DEFAULT_TS_COWORK_RUNTIME_ROLLOUT.scheduler),
-    swarm: booleanValue(rollout?.swarm, DEFAULT_TS_COWORK_RUNTIME_ROLLOUT.swarm),
-  };
-}
-
-function asRecord(value: unknown): Record<string, unknown> | undefined {
-  return value !== null && typeof value === "object" && !Array.isArray(value) ? value as Record<string, unknown> : undefined;
-}
-
-function booleanValue(value: unknown, fallback: boolean): boolean {
-  return typeof value === "boolean" ? value : fallback;
-}
-
-function abortAfter(controller: AbortController, timeoutMs: number, label: string): ReturnType<typeof globalThis.setTimeout> {
-  return globalThis.setTimeout(() => {
-    controller.abort(new Error(`${label} timed out after ${timeoutMs} ms`));
-  }, timeoutMs);
-}
-
-export type NativeSkillsApi = {
-  list: () => Promise<unknown>;
-  detail: (name: string) => Promise<unknown>;
-  create: (body: unknown) => Promise<unknown>;
-  update: (name: string, body: unknown) => Promise<unknown>;
-  delete: (name: string) => Promise<unknown>;
-  validate: (name: string) => Promise<unknown>;
-};
-
-export type NativeSessionsApi = {
-  list: () => Promise<unknown>;
-  messages: (key: string) => Promise<unknown>;
-  effectiveCapabilities?: (key: string) => Promise<unknown>;
-  turns?: (key: string) => Promise<unknown>;
-  agentTurnRuntimeState?: (key: string, turnId: string) => Promise<unknown>;
-  delete?: (key: string) => Promise<unknown>;
-  patch?: (key: string, body: unknown) => Promise<unknown>;
-  branch?: (body: unknown) => Promise<unknown>;
-  clear?: (key: string) => Promise<unknown>;
-  upsertTaskProgress?: (key: string, body: unknown) => Promise<unknown>;
-};
-
-export type NativeThreadsApi = {
-  create: (body?: unknown) => Promise<unknown>;
-  read: (body: unknown) => Promise<unknown>;
-  resume: (body: unknown) => Promise<unknown>;
-  list: (body?: unknown) => Promise<unknown>;
-  search: (body: unknown) => Promise<unknown>;
-  activity: (body: unknown) => Promise<unknown>;
-  status: (body: unknown) => Promise<unknown>;
-  updateMetadata: (body: unknown) => Promise<unknown>;
-  agentRegistry: (body?: unknown) => Promise<unknown>;
-  startTurn: (body: unknown) => Promise<unknown>;
-  continueTurn: (body: unknown) => Promise<unknown>;
-  interrupt: (body: unknown) => Promise<unknown>;
-  applyOp: (body: unknown) => Promise<unknown>;
-  archive: (body: unknown) => Promise<unknown>;
-  unarchive: (body: unknown) => Promise<unknown>;
-  delete: (body: unknown) => Promise<unknown>;
-  fork: (body: unknown) => Promise<unknown>;
-  events: (body: unknown) => Promise<unknown>;
-  restoreCheckpoint: (body: unknown) => Promise<unknown>;
-  submitTurn: (body: unknown) => Promise<unknown>;
-  submitForm: (body: unknown) => Promise<unknown>;
-};
-
-export type NativeConfigApi = {
-  get: () => Promise<unknown>;
-};
-
-export type NativeWorkspaceApi = {
-  files: () => Promise<unknown>;
-  file: (path: string) => Promise<unknown>;
-  putFile: (path: string, body: unknown) => Promise<unknown>;
-  directory: (request: { cursor?: string; nameQuery?: string; path: string }) => Promise<unknown>;
-  fileChunk: (request: { cursor?: string; path: string }) => Promise<unknown>;
-};
-
-export type NativeCoworkRouteRequest = {
-  method: string;
-  path: string;
-  query?: Record<string, string>;
-  body?: unknown;
-};
-
-export type NativeCoworkApi = {
-  route: (request: NativeCoworkRouteRequest) => Promise<unknown>;
-};
-
-export type NativeWebuiRouteRequest = {
-  method: string;
-  path: string;
-  headers?: Record<string, unknown>;
-  body?: unknown;
-};
-
-export type NativeWebuiRouteResponse = {
-  status: number;
-  body?: unknown;
-  headers?: Record<string, unknown>;
-};
-
-export type NativeWebuiApi = {
-  route: (request: NativeWebuiRouteRequest) => Promise<unknown>;
-  routeResponse?: (request: NativeWebuiRouteRequest) => Promise<NativeWebuiRouteResponse>;
-};
-
-export type WebuiApprovalsListOptions = {
-  sessionKey?: string;
-  chatId?: string;
-  channel?: string;
-};
-
-type WebSocketProbe = (url: string, timeoutMs: number) => Promise<ProbeResult>;
-
-type BootstrapSession = {
-  token: string;
-  wsPath: string;
-  refreshTokenPath: string;
-  tokenTtlS: number;
-};
-
-type TrackedSession = BootstrapSession & {
-  expiresAtMs: number;
-};
-
-type ProbeResult =
-  | {
-      ok: true;
-    }
-  | {
-      ok: false;
-      error: string;
-    };
-
-export type GatewayHealth = {
-  state: "running" | "degraded" | "offline";
-  httpBaseUrl: string;
-  wsUrl: string;
-  tokenReady: boolean;
-  http: ProbeResult & {
-    payload?: unknown;
-  };
-  webSocket: ProbeResult;
-  checkedAt: string;
-};
-
-type HealthOptions = {
-  config?: GatewayConfig;
-  fetchFn?: FetchFn;
-  webSocketProbe?: WebSocketProbe;
-};
-
-const TOKEN_REFRESH_MARGIN_MS = 60_000;
-
-export async function checkGatewayHealth(options: HealthOptions = {}): Promise<GatewayHealth> {
-  const config = options.config ?? DEFAULT_GATEWAY_CONFIG;
-  const fetchFn = options.fetchFn ?? fetch;
-  const bootstrap = await bootstrapGateway(config, fetchFn);
-  const http = bootstrap.ok
-    ? await checkHttpStatus(config, fetchFn, bootstrap.session.token)
-    : { ok: false as const, error: bootstrap.error };
-  const wsUrl = bootstrap.ok ? authenticatedWsUrl(config, bootstrap.session) : config.wsUrl;
-  const webSocket =
-    http.ok && options.webSocketProbe
-      ? await options.webSocketProbe(wsUrl, config.requestTimeoutMs)
-      : http.ok
-        ? await probeWebSocket(wsUrl, config.requestTimeoutMs)
-        : { ok: false as const, error: "HTTP status unavailable" };
-
-  return {
-    state: http.ok && webSocket.ok ? "running" : http.ok ? "degraded" : "offline",
-    httpBaseUrl: config.httpBaseUrl,
-    wsUrl,
-    tokenReady: bootstrap.ok,
-    http,
-    webSocket,
-    checkedAt: new Date().toISOString(),
-  };
-}
-
-async function checkHttpStatus(
-  config: GatewayConfig,
-  fetchFn: FetchFn,
-  token: string,
-): Promise<GatewayHealth["http"]> {
-  const controller = new AbortController();
-  const timeout = abortAfter(controller, config.requestTimeoutMs, "Gateway status request");
-  try {
-    const response = await fetchFn(`${config.httpBaseUrl}/api/status`, {
-      headers: authHeaders(token),
-      signal: controller.signal,
-    });
-    if (!response.ok) {
-      return { ok: false, error: `HTTP ${response.status}` };
-    }
-    return { ok: true, payload: await response.json().catch(() => null) };
-  } catch (error) {
-    return { ok: false, error: stringifyError(error) };
-  } finally {
-    globalThis.clearTimeout(timeout);
-  }
-}
-
-export async function probeWebSocket(url: string, timeoutMs: number): Promise<ProbeResult> {
-  return new Promise((resolve) => {
-    const socket = new WebSocket(url);
-    const timeout = globalThis.setTimeout(() => {
-      socket.close();
-      resolve({ ok: false, error: "WebSocket probe timed out" });
-    }, timeoutMs);
-
-    socket.addEventListener(
-      "open",
-      () => {
-        globalThis.clearTimeout(timeout);
-        socket.close();
-        resolve({ ok: true });
-      },
-      { once: true },
-    );
-    socket.addEventListener(
-      "error",
-      () => {
-        globalThis.clearTimeout(timeout);
-        resolve({ ok: false, error: "WebSocket connection failed" });
-      },
-      { once: true },
-    );
-  });
-}
-
-export function createGatewayApiClient(options: ClientOptions = {}) {
-  const config = options.config ?? DEFAULT_GATEWAY_CONFIG;
-  const fetchFn = options.fetchFn ?? fetch;
-  let sessionPromise: Promise<TrackedSession> | null = null;
-  let refreshPromise: Promise<TrackedSession> | null = null;
-  const startBootstrap = () =>
-    bootstrapGateway(config, fetchFn).then((result) => {
-      if (!result.ok) {
-        throw new Error(`Gateway bootstrap failed: ${result.error}`);
-      }
-      return trackSession(result.session);
-    });
-  const createBootstrapSessionPromise = () => {
-    const next = startBootstrap().catch((error) => {
-      if (sessionPromise === next) {
-        sessionPromise = null;
-      }
-      throw error;
-    });
-    return next;
-  };
-  const getSession = async (sessionOptions: { forceBootstrap?: boolean } = {}) => {
-    if (sessionOptions.forceBootstrap) {
-      sessionPromise = createBootstrapSessionPromise();
-      return sessionPromise;
-    }
-    sessionPromise ??= createBootstrapSessionPromise();
-    const session = await sessionPromise;
-    if (!shouldRefreshSession(session)) {
-      return session;
-    }
-    refreshPromise ??= refreshGateway(config, fetchFn, session, options.nativeWebui)
-      .then((result) => {
-        if (result.ok) {
-          return trackSession(result.session);
-        }
-        const nextSessionPromise = createBootstrapSessionPromise();
-        sessionPromise = nextSessionPromise;
-        return nextSessionPromise;
-      })
-      .then((nextSession) => {
-        sessionPromise = Promise.resolve(nextSession);
-        return nextSession;
-      })
-      .finally(() => {
-        refreshPromise = null;
-      });
-    return refreshPromise;
-  };
-  const request = async (path: string, init?: RequestInit) => {
-    const session = await getSession();
-    try {
-      return await requestJson(config, fetchFn, path, withAuth(init, session.token));
-    } catch (error) {
-      if (!isGatewayUnauthorizedError(error)) {
-        throw error;
-      }
-      const freshSession = await getSession({ forceBootstrap: true });
-      return requestJson(config, fetchFn, path, withAuth(init, freshSession.token));
-    }
-  };
-
-  return {
-    runtime: {
-      status: () => nativeOrGateway(
-        () => options.nativeWebui?.route({ method: "GET", path: "/api/status" }),
-        () => request("/api/status"),
-        "webui.status",
-      ),
-    },
-    openAi: {
-      health: () => nativeOrGateway(
-        () => options.nativeWebui?.route({ method: "GET", path: "/health" }),
-        () => request("/health"),
-        "openai.health",
-      ),
-      models: () => nativeOrGateway(
-        () => options.nativeWebui?.route({ method: "GET", path: "/v1/models" }),
-        () => request("/v1/models"),
-        "openai.models",
-      ),
-    },
-    sessions: {
-      list: () => nativeOrGateway(
-        () => options.nativeSessions?.list() ?? options.nativeWebui?.route({ method: "GET", path: "/api/sessions" }),
-        () => request("/api/sessions"),
-        "webui.sessions.list",
-        !options.nativeSessions,
-      ),
-      messages: (key: string) => nativeOrGateway(
-        () => options.nativeSessions?.messages(key) ?? options.nativeWebui?.route({ method: "GET", path: `/api/sessions/${encodePathSegment(key)}/messages` }),
-        () => request(`/api/sessions/${encodePathSegment(key)}/messages`),
-        "webui.sessions.messages",
-        !options.nativeSessions,
-      ),
-      effectiveCapabilities: (key: string) => nativeOrGateway(
-        () => options.nativeSessions?.effectiveCapabilities?.(key) ?? options.nativeWebui?.route({
-          method: "GET",
-          path: `/api/sessions/${encodePathSegment(key)}/effective-capabilities`,
-        }),
-        () => request(`/api/sessions/${encodePathSegment(key)}/effective-capabilities`),
-        "webui.sessions.effectiveCapabilities",
-        !options.nativeSessions?.effectiveCapabilities,
-      ),
-      turns: (key: string) => nativeOrGateway(
-        () => options.nativeSessions?.turns?.(key),
-        () => Promise.resolve({ sessionId: key, turns: [] }),
-        "webui.sessions.turns",
-        !options.nativeSessions?.turns,
-      ),
-      agentTurnRuntimeState: (key: string, turnId: string) => nativeOrGateway(
-        () => options.nativeSessions?.agentTurnRuntimeState?.(key, turnId),
-        () => Promise.resolve(null),
-        "webui.sessions.agentTurnRuntimeState",
-        !options.nativeSessions?.agentTurnRuntimeState,
-      ),
-      profile: (key: string) => nativeOrGateway(
-        () => options.nativeWebui?.route({ method: "GET", path: `/api/sessions/${encodePathSegment(key)}/profile` }),
-        () => request(`/api/sessions/${encodePathSegment(key)}/profile`),
-        "webui.sessions.profile",
-      ),
-      delete: (key: string) => nativeOrGateway(
-        () => options.nativeSessions?.delete?.(key) ?? options.nativeWebui?.route({ method: "DELETE", path: `/api/sessions/${encodePathSegment(key)}` }),
-        () => request(`/api/sessions/${encodePathSegment(key)}`, { method: "DELETE" }),
-        "webui.sessions.delete",
-        !options.nativeSessions?.delete,
-      ),
-      patch: (key: string, body: unknown) => nativeOrGateway(
-        () => options.nativeSessions?.patch?.(key, body) ?? options.nativeWebui?.route({
-          method: "PATCH",
-          path: `/api/sessions/${encodePathSegment(key)}`,
-          body,
-        }),
-        () => request(`/api/sessions/${encodePathSegment(key)}`, jsonRequest("PATCH", body)),
-        "webui.sessions.patch",
-        !options.nativeSessions?.patch,
-      ),
-      branch: (body: unknown) => nativeOrGateway(
-        () => options.nativeSessions?.branch?.(body) ?? options.nativeWebui?.route({
-          method: "POST",
-          path: "/api/sessions/branch",
-          body,
-        }),
-        () => request("/api/sessions/branch", jsonRequest("POST", body)),
-        "webui.sessions.branch",
-        !options.nativeSessions?.branch,
-      ),
-      clear: (key: string) => nativeOrGateway(
-        () => options.nativeSessions?.clear?.(key) ?? options.nativeWebui?.route({
-          method: "POST",
-          path: `/api/sessions/${encodePathSegment(key)}/clear`,
-        }),
-        () => request(`/api/sessions/${encodePathSegment(key)}/clear`, { method: "POST" }),
-        "webui.sessions.clear",
-        !options.nativeSessions?.clear,
-      ),
-    },
-    threads: {
-      create: (body?: unknown) => nativeThreadRequest(options.nativeThreads?.create(body), "threads.create"),
-      read: (body: unknown) => nativeThreadRequest(options.nativeThreads?.read(body), "threads.read"),
-      resume: (body: unknown) => nativeThreadRequest(options.nativeThreads?.resume(body), "threads.resume"),
-      list: (body?: unknown) => nativeThreadRequest(options.nativeThreads?.list(body), "threads.list"),
-      search: (body: unknown) => nativeThreadRequest(options.nativeThreads?.search(body), "threads.search"),
-      activity: (body: unknown) => nativeThreadRequest(options.nativeThreads?.activity(body), "threads.activity"),
-      status: (body: unknown) => nativeThreadRequest(options.nativeThreads?.status(body), "threads.status"),
-      updateMetadata: (body: unknown) =>
-        nativeThreadRequest(options.nativeThreads?.updateMetadata(body), "threads.updateMetadata"),
-      agentRegistry: (body?: unknown) =>
-        nativeThreadRequest(options.nativeThreads?.agentRegistry(body), "threads.agentRegistry"),
-      startTurn: (body: unknown) => nativeThreadRequest(options.nativeThreads?.startTurn(body), "threads.startTurn"),
-      continueTurn: (body: unknown) =>
-        nativeThreadRequest(options.nativeThreads?.continueTurn(body), "threads.continueTurn"),
-      interrupt: (body: unknown) => nativeThreadRequest(options.nativeThreads?.interrupt(body), "threads.interrupt"),
-      applyOp: (body: unknown) => nativeThreadRequest(options.nativeThreads?.applyOp(body), "threads.applyOp"),
-      archive: (body: unknown) => nativeThreadRequest(options.nativeThreads?.archive(body), "threads.archive"),
-      unarchive: (body: unknown) => nativeThreadRequest(options.nativeThreads?.unarchive(body), "threads.unarchive"),
-      delete: (body: unknown) => nativeThreadRequest(options.nativeThreads?.delete(body), "threads.delete"),
-      fork: (body: unknown) => nativeThreadRequest(options.nativeThreads?.fork(body), "threads.fork"),
-      events: (body: unknown) => nativeThreadRequest(options.nativeThreads?.events(body), "threads.events"),
-      restoreCheckpoint: (body: unknown) =>
-        nativeThreadRequest(options.nativeThreads?.restoreCheckpoint(body), "threads.restoreCheckpoint"),
-      submitTurn: (body: unknown) => nativeThreadRequest(options.nativeThreads?.submitTurn(body), "threads.submitTurn"),
-      submitForm: (body: unknown) => nativeThreadRequest(options.nativeThreads?.submitForm(body), "threads.submitForm"),
-    },
-    config: {
-      get: () => nativeOrGateway(
-        () => options.nativeConfig?.get() ?? options.nativeWebui?.route({ method: "GET", path: "/api/config" }),
-        () => request("/api/config"),
-        "webui.config.get",
-      ),
-      patch: (body: unknown) => nativeOrGateway(
-        () => options.nativeWebui?.route({ method: "PATCH", path: "/api/config", body }),
-        () => request("/api/config", jsonRequest("PATCH", body)),
-        "webui.config.patch",
-      ),
-      providers: () => nativeOrGateway(
-        () => options.nativeWebui?.route({ method: "GET", path: "/api/providers" }),
-        () => request("/api/providers"),
-        "webui.providers",
-      ),
-      providerModels: (body: unknown) => nativeOrGateway(
-        () => options.nativeWebui?.route({ method: "POST", path: "/api/provider-models", body }),
-        () => request("/api/provider-models", jsonRequest("POST", body)),
-        "webui.providerModels",
-      ),
-    },
-    tools: {
-      list: () => nativeOrGateway(
-        () => options.nativeWebui?.route({ method: "GET", path: "/api/tools" }),
-        () => request("/api/tools"),
-        "webui.tools.list",
-      ),
-      approvals: (approvalOptions?: WebuiApprovalsListOptions) => {
-        const path = approvalsListPath(approvalOptions);
-        return request(path);
-      },
-      approveApproval: (approvalId: string, body: unknown) => {
-        const path = `/api/approvals/${encodePathSegment(approvalId)}/approve`;
-        return request(path, jsonRequest("POST", body));
-      },
-      denyApproval: (approvalId: string, body: unknown) => {
-        const path = `/api/approvals/${encodePathSegment(approvalId)}/deny`;
-        return request(path, jsonRequest("POST", body));
-      },
-    },
-    skills: {
-      list: () => nativeOrGateway(
-        () => options.nativeSkills?.list() ?? options.nativeWebui?.route({ method: "GET", path: "/api/skills" }),
-        () => request("/api/skills"),
-        "skills.list",
-      ),
-      detail: (name: string) => nativeOrGateway(
-        () => options.nativeSkills?.detail(name) ?? options.nativeWebui?.route({
-          method: "GET",
-          path: `/api/skills/${encodePathSegment(name)}`,
-        }),
-        () => request(`/api/skills/${encodePathSegment(name)}`),
-        "skills.detail",
-      ),
-      create: (body: unknown) => nativeOrGateway(
-        () => options.nativeSkills?.create(body)
-          ?? options.nativeWebui?.route({ method: "POST", path: "/api/skills", body }),
-        () => request("/api/skills", jsonRequest("POST", body)),
-        "skills.create",
-      ),
-      update: (name: string, body: unknown) =>
-        nativeOrGateway(
-          () => options.nativeSkills?.update(name, body) ?? options.nativeWebui?.route({
-            method: "PATCH",
-            path: `/api/skills/${encodePathSegment(name)}`,
-            body,
-          }),
-          () => request(`/api/skills/${encodePathSegment(name)}`, jsonRequest("PATCH", body)),
-          "skills.update",
-        ),
-      delete: (name: string) => nativeOrGateway(
-        () => options.nativeSkills?.delete(name) ?? options.nativeWebui?.route({
-          method: "DELETE",
-          path: `/api/skills/${encodePathSegment(name)}`,
-        }),
-        () => request(`/api/skills/${encodePathSegment(name)}`, { method: "DELETE" }),
-        "skills.delete",
-      ),
-      validate: (name: string) => nativeOrGateway(
-        () => options.nativeSkills?.validate(name) ?? options.nativeWebui?.route({
-          method: "POST",
-          path: `/api/skills/${encodePathSegment(name)}/validate`,
-        }),
-        () => request(`/api/skills/${encodePathSegment(name)}/validate`, { method: "POST" }),
-        "skills.validate",
-      ),
-    },
-    agentUi: {
-      submitForm: (formId: string, body: unknown) => {
-        const path = `/api/agent-ui/forms/${encodePathSegment(formId)}/submit`;
-        return nativeOrGateway(
-          () => options.nativeWebui?.route({ method: "POST", path, body }),
-          () => request(path, jsonRequest("POST", body)),
-          "webui.agentUi.submitForm",
-        );
-      },
-      cancelForm: (formId: string, body: unknown) => {
-        const path = `/api/agent-ui/forms/${encodePathSegment(formId)}/cancel`;
-        return nativeOrGateway(
-          () => options.nativeWebui?.route({ method: "POST", path, body }),
-          () => request(path, jsonRequest("POST", body)),
-          "webui.agentUi.cancelForm",
-        );
-      },
-    },
-    workspace: {
-      files: () => nativeOrGateway(
-        () => options.nativeWorkspace?.files() ?? options.nativeWebui?.route({ method: "GET", path: "/api/workspace/files" }),
-        () => request("/api/workspace/files"),
-        "webui.workspace.files",
-      ),
-      file: (path: string) => {
-        const routePath = `/api/workspace/files/${encodePathSegment(path)}`;
-        return nativeOrGateway(
-          () => options.nativeWorkspace?.file(path) ?? options.nativeWebui?.route({ method: "GET", path: routePath }),
-          () => request(routePath),
-          "webui.workspace.file",
-        );
-      },
-      putFile: (path: string, body: unknown) => {
-        const routePath = `/api/workspace/files/${encodePathSegment(path)}`;
-        return nativeOrGateway(
-          () => options.nativeWorkspace?.putFile(path, body) ?? options.nativeWebui?.route({ method: "PUT", path: routePath, body }),
-          () => request(routePath, jsonRequest("PUT", body)),
-          "webui.workspace.putFile",
-        );
-      },
-      directory: (directoryRequest: { cursor?: string; nameQuery?: string; path: string }) => {
-        const params = new URLSearchParams({ path: directoryRequest.path });
-        if (directoryRequest.cursor) params.set("cursor", directoryRequest.cursor);
-        if (directoryRequest.nameQuery) params.set("nameQuery", directoryRequest.nameQuery);
-        const routePath = `/api/workspace/directory?${params.toString()}`;
-        return nativeOrGateway(
-          () => options.nativeWorkspace?.directory(directoryRequest)
-            ?? options.nativeWebui?.route({ method: "GET", path: routePath }),
-          () => request(routePath),
-          "webui.workspace.directory",
-        );
-      },
-      fileChunk: (fileRequest: { cursor?: string; path: string }) => {
-        const params = new URLSearchParams({ path: fileRequest.path });
-        if (fileRequest.cursor) params.set("cursor", fileRequest.cursor);
-        const routePath = `/api/workspace/read?${params.toString()}`;
-        return nativeOrGateway(
-          () => options.nativeWorkspace?.fileChunk(fileRequest)
-            ?? options.nativeWebui?.route({ method: "GET", path: routePath }),
-          () => request(routePath),
-          "webui.workspace.read",
-        );
-      },
-    },
-    cowork: {
-      sessions: (sessionOptions: { includeCompleted?: boolean; originChatId?: string } = {}) => {
-        const params = new URLSearchParams();
-        if (sessionOptions.includeCompleted) {
-          params.set("include_completed", "true");
-        }
-        if (sessionOptions.originChatId) {
-          params.set("origin_chat_id", sessionOptions.originChatId);
-        }
-        const path = `/api/cowork/sessions${params.toString() ? `?${params}` : ""}`;
-        return coworkNativeOrGateway(
-          options.nativeCowork,
-          options.tsCoworkRuntime,
-          request,
-          "GET",
-          path,
-          undefined,
-          "cowork.sessions",
-        );
-      },
-      session: (sessionId: string) => coworkNativeOrGateway(
-        options.nativeCowork,
-        options.tsCoworkRuntime,
-        request,
-        "GET",
-        `/api/cowork/sessions/${encodePathSegment(sessionId)}`,
-        undefined,
-        "cowork.session",
-      ),
-      summary: (sessionId: string) => coworkNativeOrGateway(
-        options.nativeCowork,
-        options.tsCoworkRuntime,
-        request,
-        "GET",
-        `/api/cowork/sessions/${encodePathSegment(sessionId)}/summary`,
-        undefined,
-        "cowork.summary",
-      ),
-      graph: (sessionId: string) => coworkNativeOrGateway(
-        options.nativeCowork,
-        options.tsCoworkRuntime,
-        request,
-        "GET",
-        `/api/cowork/sessions/${encodePathSegment(sessionId)}/graph`,
-        undefined,
-        "cowork.graph",
-      ),
-      agentActivity: (sessionId: string, agentId: string, activityOptions: { limit?: number } = {}) => {
-        const params = new URLSearchParams();
-        if (activityOptions.limit !== undefined) {
-          params.set("limit", String(activityOptions.limit));
-        }
-        const path = `/api/cowork/sessions/${encodePathSegment(sessionId)}/agents/${encodePathSegment(agentId)}/activity${params.toString() ? `?${params}` : ""}`;
-        return coworkNativeOrGateway(
-          options.nativeCowork,
-          options.tsCoworkRuntime,
-          request,
-          "GET",
-          path,
-          undefined,
-          "cowork.agentActivity",
-        );
-      },
-      observation: (sessionId: string, detailRef: string, observationOptions: { requesterAgentId?: string } = {}) => {
-        const params = new URLSearchParams();
-        if (observationOptions.requesterAgentId) {
-          params.set("agent_id", observationOptions.requesterAgentId);
-        }
-        const path = `/api/cowork/sessions/${encodePathSegment(sessionId)}/observations/${encodePathSegment(detailRef)}${params.toString() ? `?${params}` : ""}`;
-        return coworkNativeOrGateway(
-          options.nativeCowork,
-          options.tsCoworkRuntime,
-          request,
-          "GET",
-          path,
-          undefined,
-          "cowork.observation",
-        );
-      },
-      blueprint: (sessionId: string) => coworkNativeOrGateway(
-        options.nativeCowork,
-        options.tsCoworkRuntime,
-        request,
-        "GET",
-        `/api/cowork/sessions/${encodePathSegment(sessionId)}/blueprint`,
-        undefined,
-        "cowork.blueprint",
-      ),
-      trace: (sessionId: string) => coworkNativeOrGateway(
-        options.nativeCowork,
-        options.tsCoworkRuntime,
-        request,
-        "GET",
-        `/api/cowork/sessions/${encodePathSegment(sessionId)}/trace`,
-        undefined,
-        "cowork.trace",
-      ),
-      dag: (sessionId: string) => coworkNativeOrGateway(
-        options.nativeCowork,
-        options.tsCoworkRuntime,
-        request,
-        "GET",
-        `/api/cowork/sessions/${encodePathSegment(sessionId)}/dag`,
-        undefined,
-        "cowork.dag",
-      ),
-      artifacts: (sessionId: string) => coworkNativeOrGateway(
-        options.nativeCowork,
-        options.tsCoworkRuntime,
-        request,
-        "GET",
-        `/api/cowork/sessions/${encodePathSegment(sessionId)}/artifacts`,
-        undefined,
-        "cowork.artifacts",
-      ),
-      organization: (sessionId: string) => coworkNativeOrGateway(
-        options.nativeCowork,
-        options.tsCoworkRuntime,
-        request,
-        "GET",
-        `/api/cowork/sessions/${encodePathSegment(sessionId)}/organization`,
-        undefined,
-        "cowork.organization",
-      ),
-      queues: (sessionId: string) => coworkNativeOrGateway(
-        options.nativeCowork,
-        options.tsCoworkRuntime,
-        request,
-        "GET",
-        `/api/cowork/sessions/${encodePathSegment(sessionId)}/queues`,
-        undefined,
-        "cowork.queues",
-      ),
-      branches: (sessionId: string) => coworkNativeOrGateway(
-        options.nativeCowork,
-        options.tsCoworkRuntime,
-        request,
-        "GET",
-        `/api/cowork/sessions/${encodePathSegment(sessionId)}/branches`,
-        undefined,
-        "cowork.branches",
-      ),
-      create: (body: unknown) => coworkNativeOrGateway(
-        options.nativeCowork,
-        options.tsCoworkRuntime,
-        request,
-        "POST",
-        "/api/cowork/sessions",
-        body,
-        "cowork.create",
-      ),
-      run: (sessionId: string, body: unknown) => coworkNativeOrGateway(
-        options.nativeCowork,
-        options.tsCoworkRuntime,
-        request,
-        "POST",
-        `/api/cowork/sessions/${encodePathSegment(sessionId)}/run`,
-        body,
-        "cowork.run",
-      ),
-      updateBudget: (sessionId: string, body: unknown, budgetOptions: { method?: "POST" | "PATCH" } = {}) => coworkNativeOrGateway(
-        options.nativeCowork,
-        options.tsCoworkRuntime,
-        request,
-        budgetOptions.method ?? "POST",
-        `/api/cowork/sessions/${encodePathSegment(sessionId)}/budget`,
-        body,
-        "cowork.updateBudget",
-      ),
-      action: (sessionId: string, action: "pause" | "resume" | "emergency-stop", body?: unknown) => coworkNativeOrGateway(
-        options.nativeCowork,
-        options.tsCoworkRuntime,
-        request,
-        "POST",
-        `/api/cowork/sessions/${encodePathSegment(sessionId)}/${action}`,
-        body,
-        `cowork.${action}`,
-      ),
-      delete: (sessionId: string) => coworkNativeOrGateway(
-        options.nativeCowork,
-        options.tsCoworkRuntime,
-        request,
-        "DELETE",
-        `/api/cowork/sessions/${encodePathSegment(sessionId)}`,
-        undefined,
-        "cowork.delete",
-      ),
-      message: (sessionId: string, body: unknown) => coworkNativeOrGateway(
-        options.nativeCowork,
-        options.tsCoworkRuntime,
-        request,
-        "POST",
-        `/api/cowork/sessions/${encodePathSegment(sessionId)}/messages`,
-        body,
-        "cowork.message",
-      ),
-      addTask: (sessionId: string, body: unknown) => coworkNativeOrGateway(
-        options.nativeCowork,
-        options.tsCoworkRuntime,
-        request,
-        "POST",
-        `/api/cowork/sessions/${encodePathSegment(sessionId)}/tasks`,
-        body,
-        "cowork.addTask",
-      ),
-      taskAction: (sessionId: string, taskId: string, action: "assign" | "retry" | "review", body: unknown = {}) => coworkNativeOrGateway(
-        options.nativeCowork,
-        options.tsCoworkRuntime,
-        request,
-        "POST",
-        `/api/cowork/sessions/${encodePathSegment(sessionId)}/tasks/${encodePathSegment(taskId)}/${action}`,
-        body,
-        `cowork.task.${action}`,
-      ),
-      workUnitAction: (sessionId: string, workUnitId: string, action: "retry" | "skip" | "cancel", body: unknown = {}) => coworkNativeOrGateway(
-        options.nativeCowork,
-        options.tsCoworkRuntime,
-        request,
-        "POST",
-        `/api/cowork/sessions/${encodePathSegment(sessionId)}/work-units/${encodePathSegment(workUnitId)}/${action}`,
-        body,
-        `cowork.workUnit.${action}`,
-      ),
-      selectBranch: (sessionId: string, branchId: string, body: unknown = undefined) => coworkNativeOrGateway(
-        options.nativeCowork,
-        options.tsCoworkRuntime,
-        request,
-        "POST",
-        `/api/cowork/sessions/${encodePathSegment(sessionId)}/branches/${encodePathSegment(branchId)}/select`,
-        body,
-        "cowork.selectBranch",
-      ),
-      deriveBranch: (sessionId: string, sourceBranchId: string | null, body: unknown) => coworkNativeOrGateway(
-        options.nativeCowork,
-        options.tsCoworkRuntime,
-        request,
-        "POST",
-        sourceBranchId
-          ? `/api/cowork/sessions/${encodePathSegment(sessionId)}/branches/${encodePathSegment(sourceBranchId)}/derive`
-          : `/api/cowork/sessions/${encodePathSegment(sessionId)}/branches/derive`,
-        body,
-        "cowork.deriveBranch",
-      ),
-      selectBranchResult: (sessionId: string, branchId: string, body: unknown) => coworkNativeOrGateway(
-        options.nativeCowork,
-        options.tsCoworkRuntime,
-        request,
-        "POST",
-        `/api/cowork/sessions/${encodePathSegment(sessionId)}/branches/${encodePathSegment(branchId)}/result/select-final`,
-        body,
-        "cowork.selectBranchResult",
-      ),
-      mergeBranchResults: (sessionId: string, body: unknown) => coworkNativeOrGateway(
-        options.nativeCowork,
-        options.tsCoworkRuntime,
-        request,
-        "POST",
-        `/api/cowork/sessions/${encodePathSegment(sessionId)}/branch-results/merge`,
-        body,
-        "cowork.mergeBranchResults",
-      ),
-      selectFinalResult: (sessionId: string, body: unknown) => coworkNativeOrGateway(
-        options.nativeCowork,
-        options.tsCoworkRuntime,
-        request,
-        "POST",
-        `/api/cowork/sessions/${encodePathSegment(sessionId)}/final-result/select`,
-        body,
-        "cowork.selectFinalResult",
-      ),
-      mergeFinalResult: (sessionId: string, body: unknown) => coworkNativeOrGateway(
-        options.nativeCowork,
-        options.tsCoworkRuntime,
-        request,
-        "POST",
-        `/api/cowork/sessions/${encodePathSegment(sessionId)}/final-result/merge`,
-        body,
-        "cowork.mergeFinalResult",
-      ),
-      validateBlueprint: (body: unknown, validateOptions: { preview?: boolean } = {}) => coworkNativeOrGateway(
-        options.nativeCowork,
-        options.tsCoworkRuntime,
-        request,
-        "POST",
-        `/api/cowork/blueprints/${validateOptions.preview ? "preview" : "validate"}`,
-        body,
-        "cowork.validateBlueprint",
-      ),
-    },
-  };
-}
-
-function coworkNativeOrGateway(
-  nativeCowork: NativeCoworkApi | undefined,
-  _rollout: TsCoworkRuntimeRollout | undefined,
-  request: (path: string, init?: RequestInit) => Promise<unknown>,
-  method: string,
-  path: string,
-  body: unknown,
-  label: string,
-): Promise<unknown> {
-  const nativeRequest = nativeCoworkRouteRequest(method, path, body);
-  const gatewayInit = method === "GET"
-    ? undefined
-    : body === undefined
-      ? { method }
-      : jsonRequest(method, body);
-  if (!nativeCowork) {
-    return request(path, gatewayInit);
-  }
-  return nativeOrGateway(
-    () => nativeCowork.route(nativeRequest),
-    () => request(path, gatewayInit),
-    label,
-    false,
-  );
-}
-
-function nativeThreadRequest(request: Promise<unknown> | undefined, label: string): Promise<unknown> {
-  return nativeOrGateway(
-    () => request,
-    () => Promise.reject(new Error("Native thread API unavailable")),
-    label,
-    false,
-  );
-}
-
-function nativeCoworkRouteRequest(method: string, path: string, body: unknown): NativeCoworkRouteRequest {
-  const url = new URL(path, "http://desktop.local");
-  const query = Object.fromEntries(url.searchParams.entries());
-  return {
-    method,
-    path: url.pathname,
-    ...(Object.keys(query).length ? { query } : {}),
-    ...(body === undefined ? {} : { body }),
-  };
-}
-
-async function bootstrapGateway(
-  config: GatewayConfig,
-  fetchFn: FetchFn,
-): Promise<
-  | {
-      ok: true;
-      session: BootstrapSession;
-    }
-  | {
-      ok: false;
-      error: string;
-    }
-> {
-  const controller = new AbortController();
-  const timeout = abortAfter(controller, config.requestTimeoutMs, "Gateway bootstrap");
-  logDesktopNativeDebug("gateway.bootstrap.start", {
-    httpBaseUrl: config.httpBaseUrl,
-    timeoutMs: config.requestTimeoutMs,
-  });
-  try {
-    const response = await fetchFn(`${config.httpBaseUrl}/webui/bootstrap`, {
-      signal: controller.signal,
-    });
-    if (!response.ok) {
-      logDesktopNativeDebug("gateway.bootstrap.error", {
-        status: response.status,
-      });
-      return { ok: false, error: `HTTP ${response.status}` };
-    }
-    const payload = await response.json();
-    const token = typeof payload.token === "string" ? payload.token : "";
-    if (!token) {
-      logDesktopNativeDebug("gateway.bootstrap.error", {
-        error: "missing token",
-      });
-      return { ok: false, error: "bootstrap response missing token" };
-    }
-    logDesktopNativeDebug("gateway.bootstrap.complete", {
-      refreshTokenPath: typeof payload.refresh_token_path === "string" ? payload.refresh_token_path : "/webui/refresh-token",
-      tokenReady: true,
-      tokenTtlS: typeof payload.token_ttl_s === "number" ? payload.token_ttl_s : 300,
-      wsPath: typeof payload.ws_path === "string" ? payload.ws_path : "/ws",
-    });
-    return {
-      ok: true,
-      session: {
-        token,
-        wsPath: typeof payload.ws_path === "string" ? payload.ws_path : "/ws",
-        refreshTokenPath:
-          typeof payload.refresh_token_path === "string" ? payload.refresh_token_path : "/webui/refresh-token",
-        tokenTtlS: typeof payload.token_ttl_s === "number" ? payload.token_ttl_s : 300,
-      },
-    };
-  } catch (error) {
-    logDesktopNativeDebug("gateway.bootstrap.error", {
-      error: stringifyError(error),
-    });
-    return { ok: false, error: stringifyError(error) };
-  } finally {
-    globalThis.clearTimeout(timeout);
-  }
-}
-
-async function refreshGateway(
-  config: GatewayConfig,
-  fetchFn: FetchFn,
-  session: BootstrapSession,
-  nativeWebui?: NativeWebuiApi,
-): Promise<
-  | {
-      ok: true;
-      session: BootstrapSession;
-    }
-  | {
-      ok: false;
-      error: string;
-    }
-> {
-  const controller = new AbortController();
-  const timeout = abortAfter(controller, config.requestTimeoutMs, "Gateway refresh");
-  logDesktopNativeDebug("gateway.refresh.start", {
-    refreshTokenPath: session.refreshTokenPath,
-    timeoutMs: config.requestTimeoutMs,
-  });
-  try {
-    if (nativeWebui) {
-      try {
-        const payload = await nativeWebui.route({
-          method: "POST",
-          path: session.refreshTokenPath,
-          headers: { Authorization: `Bearer ${session.token}` },
-        });
-        const parsed = parseRefreshSessionPayload(payload, session);
-        if (!parsed.ok) {
-          logDesktopNativeDebug("gateway.refresh.nativeFallback", { error: parsed.error });
-        } else {
-          logDesktopNativeDebug("gateway.refresh.complete", {
-            refreshTokenPath: parsed.session.refreshTokenPath,
-            tokenReady: true,
-            tokenTtlS: parsed.session.tokenTtlS,
-            wsPath: parsed.session.wsPath,
-          });
-          return parsed;
-        }
-      } catch (error) {
-        logDesktopNativeDebug("gateway.refresh.nativeFallback", { error: stringifyError(error) });
-      }
-    }
-    const response = await fetchFn(`${config.httpBaseUrl}${session.refreshTokenPath}`, {
-      method: "POST",
-      headers: authHeaders(session.token),
-      signal: controller.signal,
-    });
-    if (!response.ok) {
-      logDesktopNativeDebug("gateway.refresh.error", {
-        status: response.status,
-      });
-      return { ok: false, error: `HTTP ${response.status}` };
-    }
-    const payload = await response.json();
-    const parsed = parseRefreshSessionPayload(payload, session);
-    if (!parsed.ok) {
-      logDesktopNativeDebug("gateway.refresh.error", {
-        error: parsed.error,
-      });
-      return parsed;
-    }
-    logDesktopNativeDebug("gateway.refresh.complete", {
-      refreshTokenPath: parsed.session.refreshTokenPath,
-      tokenReady: true,
-      tokenTtlS: parsed.session.tokenTtlS,
-      wsPath: parsed.session.wsPath,
-    });
-    return parsed;
-  } catch (error) {
-    logDesktopNativeDebug("gateway.refresh.error", {
-      error: stringifyError(error),
-    });
-    return { ok: false, error: stringifyError(error) };
-  } finally {
-    globalThis.clearTimeout(timeout);
-  }
-}
-
-function parseRefreshSessionPayload(
-  payload: unknown,
-  previousSession: BootstrapSession,
-): { ok: true; session: BootstrapSession } | { ok: false; error: string } {
-  const record = asRecord(payload);
-  const token = typeof record?.token === "string" ? record.token : "";
-  if (!token) {
-    return { ok: false, error: "refresh response missing token" };
-  }
-  return {
-    ok: true,
-    session: {
-      token,
-      wsPath: typeof record?.ws_path === "string" ? record.ws_path : previousSession.wsPath,
-      refreshTokenPath: typeof record?.refresh_token_path === "string"
-        ? record.refresh_token_path
-        : previousSession.refreshTokenPath,
-      tokenTtlS: typeof record?.token_ttl_s === "number" ? record.token_ttl_s : previousSession.tokenTtlS,
-    },
-  };
-}
-
-async function requestJson(config: GatewayConfig, fetchFn: FetchFn, path: string, init?: RequestInit): Promise<unknown> {
-  const method = init?.method ?? "GET";
-  logDesktopNativeDebug("gateway.http.request", {
-    method,
-    path,
-  });
-  const response = await fetchFn(`${config.httpBaseUrl}${path}`, {
-    headers: {
-      Accept: "application/json",
-      ...(init?.headers ?? {}),
-    },
-    ...init,
-  });
-  if (!response.ok) {
-    logDesktopNativeDebug("gateway.http.error", {
-      method,
-      path,
-      status: response.status,
-    });
-    throw new GatewayRequestError(response.status);
-  }
-  const payload = await response.json();
-  logDesktopNativeDebug("gateway.http.response", {
-    method,
-    path,
-    status: response.status,
-  });
-  return payload;
-}
-
-class GatewayRequestError extends Error {
-  constructor(readonly status: number) {
-    super(`Gateway request failed: HTTP ${status}`);
-  }
-}
-
-function isGatewayUnauthorizedError(error: unknown): boolean {
-  return error instanceof GatewayRequestError && error.status === 401;
-}
-
-function trackSession(session: BootstrapSession): TrackedSession {
-  return {
-    ...session,
-    expiresAtMs: Date.now() + session.tokenTtlS * 1000,
-  };
-}
-
-async function nativeOrGateway(
-  nativeRequest: () => Promise<unknown> | undefined,
-  gatewayRequest: () => Promise<unknown>,
-  label: string,
-  fallbackToGateway = true,
-): Promise<unknown> {
-  const request = nativeRequest();
-  if (!request) {
-    return gatewayRequest();
-  }
-  try {
-    return await request;
-  } catch (error) {
-    if (!fallbackToGateway) {
-      logDesktopNativeDebug(`${label}.nativeError`, { error: stringifyError(error) });
-      throw error;
-    }
-    logDesktopNativeDebug(`${label}.nativeFallback`, { error: stringifyError(error) });
-    return gatewayRequest();
-  }
-}
-
-function shouldRefreshSession(session: TrackedSession): boolean {
-  return Date.now() + TOKEN_REFRESH_MARGIN_MS >= session.expiresAtMs;
-}
-
-function authenticatedWsUrl(config: GatewayConfig, session: BootstrapSession): string {
-  const url = new URL(session.wsPath, config.httpBaseUrl);
-  url.protocol = url.protocol.replace(/^http/, "ws");
-  url.searchParams.set("token", session.token);
-  return url.toString();
-}
-
-function authHeaders(token: string): HeadersInit {
-  return {
-    Authorization: `Bearer ${token}`,
-  };
-}
-
-function withAuth(init: RequestInit = {}, token: string): RequestInit {
-  return {
-    ...init,
-    headers: {
-      Authorization: `Bearer ${token}`,
-      ...(init.headers ?? {}),
-    },
-  };
-}
-
-function jsonRequest(method: string, body: unknown): RequestInit {
-  return {
-    method,
-    headers: {
-      "Content-Type": "application/json",
-    },
-    body: JSON.stringify(body),
-  };
-}
-
-function encodePathSegment(value: string): string {
-  return encodeURIComponent(value);
-}
-
-function approvalsListPath(options: WebuiApprovalsListOptions | undefined): string {
-  const params = new URLSearchParams();
-  if (options?.sessionKey) {
-    params.set("session_key", options.sessionKey);
-  }
-  if (options?.chatId) {
-    params.set("chat_id", options.chatId);
-  }
-  if (options?.channel) {
-    params.set("channel", options.channel);
-  }
-  const query = params.toString();
-  return query ? `/api/approvals?${query}` : "/api/approvals";
-}
-
-function stringifyError(error: unknown): string {
-  return error instanceof Error ? error.message : String(error);
-}
