+++ src/app-core/native/desktopNativeSessions.ts
@@ -1,7 +1,19 @@
 import { invoke as tauriInvoke } from "@tauri-apps/api/core";
-import type { NativeSessionsApi } from "../gateway/gatewayHttpClient";
 
 type TauriInvoke = (command: string, args?: Record<string, unknown>) => Promise<unknown>;
+
+export type NativeSessionsApi = {
+  list: () => Promise<unknown>;
+  messages: (key: string) => Promise<unknown>;
+  effectiveCapabilities?: (key: string) => Promise<unknown>;
+  turns?: (key: string) => Promise<unknown>;
+  agentTurnRuntimeState?: (key: string, turnId: string) => Promise<unknown>;
+  delete?: (key: string) => Promise<unknown>;
+  patch?: (key: string, body: unknown) => Promise<unknown>;
+  branch?: (body: unknown) => Promise<unknown>;
+  clear?: (key: string) => Promise<unknown>;
+  upsertTaskProgress?: (key: string, body: unknown) => Promise<unknown>;
+};
 
 export function createDesktopNativeSessionsApi(options: { invoke?: TauriInvoke } = {}): NativeSessionsApi {
   const invoke = options.invoke ?? tauriInvoke;
